refactor: decompose the fetch_cloner impl monolith
Pure move: the single 4,100-line impl block in fetch_cloner/mod.rs is
split into four cohesive submodules — pending_ops (pending-request and
owned-operation bookkeeping), ownership (clone flows that resolve chain
ownership), subscription_updates (forwarded-sub processing, collisions,
greedy cloning, record parsing), and fetching (fetch/clone pipelines,
dedup, task builders). Moved methods gain pub(super); no logic changes.
mod.rs retains construction, discovery listeners, and shared state
(4,440 -> 716 lines).
