+++ magicblock-chainlink/src/chainlink/fetch_cloner/mod.rs
@@ -10,29 +10,18 @@ use std::{
 
 use dlp_api::{
     pda::delegation_record_pda_from_delegated_account,
-    state::{
-        discriminator::AccountDiscriminator, DelegationRecord,
-        UndelegationRequest,
-    },
+    state::{discriminator::AccountDiscriminator, UndelegationRequest},
 };
 use lru::LruCache;
 use magicblock_accounts_db::traits::AccountsBank;
 use magicblock_aml::RiskService;
 use magicblock_config::config::AllowedProgram;
-use magicblock_core::token_programs::{
-    is_ata, normalize_native_token_account_for_local_clone, EATA_PROGRAM_ID,
-};
 use magicblock_metrics::metrics::{
-    self, AccountFetchContext, AccountFetchReason, BankPrecheckOutcome,
-    BankPrecheckReason, ChainlinkCloneIntent,
-    ChainlinkCloneMaterializationOutcome, ChainlinkCloneOutcome,
-    ChainlinkCloneRemoteResult, ChainlinkCompanionFetchKind,
-    ChainlinkEmptyPlaceholderStage, ChainlinkPendingFetchLayer,
-    ChainlinkPendingFetchOutcome, Outcome,
+    self, AccountFetchContext, ChainlinkCompanionFetchKind,
 };
 use parking_lot::Mutex as PlMutex;
 use scc::HashMap;
-use solana_account::{Account, AccountSharedData, ReadableAccount};
+use solana_account::AccountSharedData;
 use solana_account_decoder_client_types::{
     UiAccountEncoding, UiDataSliceConfig,
 };
@@ -42,13 +31,10 @@ use solana_rpc_client_api::{
     config::{RpcAccountInfoConfig, RpcProgramAccountsConfig},
     filter::{Memcmp, RpcFilterType},
 };
-use solana_sdk_ids::system_program;
-use solana_signature::Signature;
 use solana_signer::Signer;
 use tokio::{
-    sync::{broadcast, mpsc, oneshot, Semaphore},
+    sync::{broadcast, mpsc, oneshot},
     task,
-    task::JoinSet,
 };
 use tracing::*;
 
@@ -57,11 +43,15 @@ pub(crate) const FETCH_CLONE_OPERATION_TIMEOUT: Duration =
 
 mod ata_projection;
 mod delegation;
+mod fetching;
+mod ownership;
 mod pending_clone_guard;
 mod pending_operation;
+mod pending_ops;
 mod pipeline;
 mod program_loader;
 mod subscription;
+mod subscription_updates;
 #[cfg(test)]
 mod tests;
 mod types;
@@ -714,3735 +704,5 @@ where
         &self,
     ) -> &Arc<RemoteAccountProvider<T, U>> {
         &self.remote_account_provider
-    }
-
-    #[cfg(test)]
-    fn has_pending_request(&self, pubkey: &Pubkey) -> bool {
-        self.pending_requests.contains(pubkey)
-    }
-
-    #[cfg(test)]
-    fn set_pending_operation_timeout(&self, timeout: Duration) {
-        self.pending_operation_timeout_ms
-            .store(timeout.as_millis() as u64, Ordering::Relaxed);
-    }
-
-    /// Returns the number of waiters currently registered for the pending
-    /// fetch+clone request keyed by `pubkey`, or `None` if no pending
-    /// request exists for that pubkey. Used by tests to deterministically
-    /// observe waiter registration without relying on fixed sleeps.
-    #[cfg(any(test, feature = "dev-context"))]
-    pub fn pending_request_waiter_count(
-        &self,
-        pubkey: &Pubkey,
-    ) -> Option<usize> {
-        self.pending_requests
-            .read(pubkey, |_, state| state.waiters.len())
-    }
-
-    /// Returns the number of waiters currently joined to the low-level
-    /// clone operation keyed by `pubkey`, or `None` if no clone is pending.
-    #[cfg(any(test, feature = "dev-context"))]
-    pub fn pending_clone_waiter_count(&self, pubkey: &Pubkey) -> Option<usize> {
-        let map = self
-            .pending_clones
-            .lock()
-            .expect("pending_clones mutex poisoned");
-        map.get(pubkey).map(Vec::len)
-    }
-
-    /// Cancels the in-flight fetch+clone owner for `pubkey`, if one exists.
-    pub fn cancel_pending(&self, pubkey: &Pubkey) {
-        self.pending_requests
-            .read(pubkey, |_, pending| pending.cancel.notify_one());
-    }
-
-    /// Cancels all in-flight fetch+clone owners.
-    pub fn cancel_all_pending(&self) {
-        self.pending_requests
-            .scan(|_pubkey, pending| pending.cancel.notify_one());
-    }
-
-    /// Check if a program is allowed to be cloned.
-    /// Returns true if:
-    /// - No allowed_programs restriction is set (None), OR
-    /// - The allowed_programs set is empty (treats empty as unrestricted), OR
-    /// - The program is in the allowed_programs set
-    fn is_program_allowed(&self, program_id: &Pubkey) -> bool {
-        match &self.allowed_programs {
-            None => true,
-            Some(allowed) => {
-                if allowed.is_empty() {
-                    true
-                } else {
-                    allowed.contains(program_id)
-                }
-            }
-        }
-    }
-
-    /// Attempt to claim ownership of a clone operation for a clone key.
-    /// Returns `CloneClaim::Owner` if this caller is the first and should
-    /// perform the clone. Returns `CloneClaim::Waiter(rx)` if another
-    /// caller already owns this clone and this caller should wait.
-    fn claim_pending_clone(&self, pubkey: Pubkey) -> CloneClaim {
-        let mut map = self
-            .pending_clones
-            .lock()
-            .expect("pending_clones mutex poisoned");
-        match map.entry(pubkey) {
-            hash_map::Entry::Vacant(entry) => {
-                entry.insert(Vec::new());
-                CloneClaim::Owner
-            }
-            hash_map::Entry::Occupied(mut entry) => {
-                let (tx, rx) = oneshot::channel();
-                entry.get_mut().push(tx);
-                CloneClaim::Waiter(rx)
-            }
-        }
-    }
-
-    /// Called by the owner when the clone operation completes.
-    /// Removes the pending entry and notifies all waiters.
-    fn finish_pending_clone(&self, pubkey: Pubkey, result: CloneCompletion) {
-        let waiters = {
-            let mut map = self
-                .pending_clones
-                .lock()
-                .expect("pending_clones mutex poisoned");
-            map.remove(&pubkey).unwrap_or_default()
-        };
-        for tx in waiters {
-            let _ = tx.send(result);
-        }
-    }
-
-    fn claim_or_join_owned_operation(
-        &self,
-        pubkey: Pubkey,
-        fetch_context: AccountFetchContext,
-    ) -> PendingClaim {
-        let generation = self.next_pending_request_generation();
-        let waiter_id = self.next_pending_waiter_id();
-        claim_or_join_pending(
-            self.pending_requests.clone(),
-            pubkey,
-            generation,
-            waiter_id,
-            Duration::from_millis(
-                self.pending_operation_timeout_ms.load(Ordering::Relaxed),
-            ),
-            fetch_context,
-            ChainlinkPendingFetchLayer::FetchCloner,
-        )
-    }
-
-    // Fetches every pubkey claimed by one dedup call in a single batched wire
-    // call, then spawns a per-key clone for each.
-    fn spawn_batched_owned_operation(
-        &self,
-        claimed: Vec<ClaimedOperation>,
-        mark_empty_set: &HashSet<Pubkey>,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-    ) {
-        if claimed.is_empty() {
-            return;
-        }
-        let this = self.clone();
-        let pending = self.pending_requests.clone();
-        let mark_empty = claimed
-            .iter()
-            .map(|op| op.pubkey)
-            .filter(|pubkey| mark_empty_set.contains(pubkey))
-            .collect::<HashSet<_>>();
-        task::spawn(async move {
-            let pubkeys =
-                claimed.iter().map(|op| op.pubkey).collect::<Vec<_>>();
-            let mark_empty_list =
-                mark_empty.iter().copied().collect::<Vec<_>>();
-            let mark_empty_ref = (!mark_empty_list.is_empty())
-                .then_some(mark_empty_list.as_slice());
-
-            // One shared wire call for the whole claim set, bounded by the
-            // earliest claim deadline so a hung fetch cannot leave the
-            // waiters of these operations blocked past their timeout.
-            let fetch_deadline = claimed
-                .iter()
-                .map(|op| op.deadline)
-                .min()
-                .expect("claimed is non-empty");
-            let fetch_result = match tokio::time::timeout_at(
-                fetch_deadline,
-                this.fetch_accounts(
-                    &pubkeys,
-                    mark_empty_ref,
-                    slot,
-                    fetch_context.clone(),
-                ),
-            )
-            .await
-            {
-                Ok(result) => result.map_err(|err| err.to_string()),
-                Err(_) => Err("account fetch deadline exceeded".to_string()),
-            };
-            let accs = match fetch_result {
-                Ok(accs) => accs,
-                Err(owner_msg) => {
-                    let now = tokio::time::Instant::now();
-                    for mut op in claimed {
-                        let failure = if op.deadline <= now {
-                            PendingFailure::TimedOut
-                        } else {
-                            PendingFailure::OwnerFailed(owner_msg.clone())
-                        };
-                        op.owner.finish(match failure {
-                            PendingFailure::Cancelled => {
-                                ChainlinkPendingFetchOutcome::OwnerCancelled
-                            }
-                            PendingFailure::OwnerFailed(_)
-                            | PendingFailure::TimedOut => {
-                                ChainlinkPendingFetchOutcome::OwnerFailed
-                            }
-                        });
-                        finish_pending(
-                            &pending,
-                            op.pubkey,
-                            op.generation,
-                            PendingTerminal::Failed(failure),
-                        );
-                    }
-                    return;
-                }
-            };
-            // Clone each key on its own so a cancel drops only that key's
-            // clone; the fetched account is handed over, never re-fetched.
-            for (op, account) in claimed.into_iter().zip(accs) {
-                let mark_empty_if_not_found = mark_empty.contains(&op.pubkey);
-                this.spawn_owned_operation(
-                    op,
-                    account,
-                    mark_empty_if_not_found,
-                    slot,
-                    fetch_context.clone(),
-                );
-            }
-        });
-    }
-
-    fn pending_terminal_owner_outcome(
-        terminal: &PendingTerminal,
-    ) -> ChainlinkPendingFetchOutcome {
-        match terminal {
-            PendingTerminal::Success(_) => {
-                ChainlinkPendingFetchOutcome::OwnerSucceeded
-            }
-            PendingTerminal::Failed(PendingFailure::Cancelled) => {
-                ChainlinkPendingFetchOutcome::OwnerCancelled
-            }
-            PendingTerminal::Failed(_) => {
-                ChainlinkPendingFetchOutcome::OwnerFailed
-            }
-        }
-    }
-
-    fn spawn_owned_operation(
-        &self,
-        op: ClaimedOperation,
-        account: RemoteAccount,
-        mark_empty_if_not_found: bool,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-    ) {
-        let this = self.clone();
-        let pending = self.pending_requests.clone();
-        task::spawn(async move {
-            let ClaimedOperation {
-                pubkey,
-                generation,
-                deadline,
-                cancel,
-                mut owner,
-            } = op;
-            let pubkeys = [pubkey];
-            let mark_empty = mark_empty_if_not_found.then_some(vec![pubkey]);
-            let mark_empty_ref = mark_empty.as_deref();
-            let work = this.clone_accounts(
-                &pubkeys,
-                vec![account],
-                mark_empty_ref,
-                slot,
-                fetch_context,
-            );
-            let terminal = tokio::select! {
-                biased;
-
-                result = tokio::time::timeout_at(deadline, work) => {
-                    match result {
-                        Ok(Ok(result)) => PendingTerminal::Success(result),
-                        Ok(Err(err)) => PendingTerminal::Failed(
-                            PendingFailure::OwnerFailed(err.to_string()),
-                        ),
-                        Err(_) => PendingTerminal::Failed(PendingFailure::TimedOut),
-                    }
-                }
-                _ = cancel.notified() => {
-                    PendingTerminal::Failed(PendingFailure::Cancelled)
-                }
-            };
-            let outcome = Self::pending_terminal_owner_outcome(&terminal);
-            owner.finish(outcome);
-            finish_pending(&pending, pubkey, generation, terminal);
-        });
-    }
-
-    fn next_pending_request_generation(&self) -> u64 {
-        self.pending_request_generation
-            .fetch_add(1, Ordering::Relaxed)
-    }
-
-    fn next_pending_waiter_id(&self) -> u64 {
-        self.pending_waiter_generation
-            .fetch_add(1, Ordering::Relaxed)
-    }
-
-    fn account_is_actively_delegated(account: &AccountSharedData) -> bool {
-        account.delegated() && !account.undelegating()
-    }
-
-    fn local_account_satisfies_clone_request(
-        &self,
-        request: &AccountCloneRequest,
-    ) -> bool {
-        let active_delegation_satisfies_request =
-            request.delegation_actions.is_empty()
-                && request.delegated_to_other.is_none()
-                && !request.needs_undelegation;
-        self.accounts_bank
-            .get_account(&request.pubkey)
-            .is_some_and(|account| {
-                let local_slot = account.remote_slot();
-                let request_slot = request.account.remote_slot();
-                (active_delegation_satisfies_request
-                    && Self::account_is_actively_delegated(&account))
-                    || local_slot > request_slot
-                    || (local_slot == request_slot
-                        && account.eq(&request.account))
-            })
-    }
-
-    fn is_empty_placeholder_account(account: &AccountSharedData) -> bool {
-        account.lamports() == 0
-            && account.data().is_empty()
-            && account.owner() == &Pubkey::default()
-            && !account.executable()
-    }
-
-    fn clone_remote_result_for_request(
-        request: &AccountCloneRequest,
-    ) -> ChainlinkCloneRemoteResult {
-        if Self::is_empty_placeholder_account(&request.account) {
-            ChainlinkCloneRemoteResult::NotFound
-        } else {
-            ChainlinkCloneRemoteResult::Found
-        }
-    }
-
-    fn clone_intent_for_request(
-        request: &AccountCloneRequest,
-    ) -> ChainlinkCloneIntent {
-        if Self::is_empty_placeholder_account(&request.account) {
-            ChainlinkCloneIntent::EmptyPlaceholder
-        } else if request.account.delegated() {
-            ChainlinkCloneIntent::DelegationRecord
-        } else if !request.delegation_actions.is_empty() {
-            ChainlinkCloneIntent::ActionDependency
-        } else {
-            ChainlinkCloneIntent::NormalAccount
-        }
-    }
-
-    fn record_account_materialization(
-        &self,
-        pubkey: &Pubkey,
-        request_slot: u64,
-        remote_result: ChainlinkCloneRemoteResult,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkCloneMaterializationOutcome {
-        let outcome = if self
-            .accounts_bank
-            .get_account(pubkey)
-            .is_some_and(|account| account.remote_slot() >= request_slot)
-        {
-            ChainlinkCloneMaterializationOutcome::ObservedInBankAfterEnsure
-        } else {
-            ChainlinkCloneMaterializationOutcome::StillMissingAfterEnsure
-        };
-        metrics::inc_chainlink_clone_materialization_accounts_total_with_context(
-            fetch_context,
-            remote_result,
-            outcome,
-        );
-        outcome
-    }
-
-    fn record_empty_placeholder_stage(
-        is_empty_placeholder: bool,
-        fetch_context: AccountFetchContext,
-        stage: ChainlinkEmptyPlaceholderStage,
-        outcome: Outcome,
-    ) {
-        if is_empty_placeholder {
-            metrics::inc_chainlink_empty_placeholder_accounts_total_with_context(
-                fetch_context,
-                stage,
-                outcome,
-            );
-        }
-    }
-
-    fn record_empty_placeholder_materialization_stage(
-        is_empty_placeholder: bool,
-        fetch_context: AccountFetchContext,
-        materialization_outcome: ChainlinkCloneMaterializationOutcome,
-    ) {
-        let (stage, outcome) = match materialization_outcome {
-            ChainlinkCloneMaterializationOutcome::ObservedInBankAfterEnsure => (
-                ChainlinkEmptyPlaceholderStage::ObservedInBankAfterEnsure,
-                Outcome::Success,
-            ),
-            ChainlinkCloneMaterializationOutcome::StillMissingAfterEnsure
-            | ChainlinkCloneMaterializationOutcome::RemovedAfterMaterialization => (
-                ChainlinkEmptyPlaceholderStage::StillMissingAfterEnsure,
-                Outcome::Error,
-            ),
-        };
-        Self::record_empty_placeholder_stage(
-            is_empty_placeholder,
-            fetch_context,
-            stage,
-            outcome,
-        );
-    }
-
-    fn record_program_materialization(
-        &self,
-        program_id: &Pubkey,
-        remote_slot: u64,
-        fetch_context: AccountFetchContext,
-    ) {
-        let outcome = if self
-            .accounts_bank
-            .get_account(program_id)
-            .is_some_and(|account| account.remote_slot() >= remote_slot)
-        {
-            ChainlinkCloneMaterializationOutcome::ObservedInBankAfterEnsure
-        } else {
-            ChainlinkCloneMaterializationOutcome::StillMissingAfterEnsure
-        };
-        metrics::inc_chainlink_clone_materialization_accounts_total_with_context(
-            fetch_context,
-            ChainlinkCloneRemoteResult::Found,
-            outcome,
-        );
-    }
-
-    /// Submits a clone request through ownership coordination.
-    /// Only one account clone per pubkey is submitted at a time. Waiters
-    /// retry local freshness checks after a successful owner so a newer
-    /// request can clone after an older request finishes.
-    async fn clone_account_with_ownership(
-        &self,
-        request: AccountCloneRequest,
-        fetch_context: AccountFetchContext,
-    ) -> ClonerResult<Signature> {
-        let pubkey = request.pubkey;
-        let remote_result = Self::clone_remote_result_for_request(&request);
-        let clone_intent = Self::clone_intent_for_request(&request);
-        let mut request = Some(request);
-
-        loop {
-            let Some(request_ref) = request.as_ref() else {
-                return Err(ClonerError::FailedToCloneRegularAccount(
-                    pubkey,
-                    Box::new(ClonerError::CommittorServiceError(
-                        "missing clone request before ownership claim"
-                            .to_string(),
-                    )),
-                ));
-            };
-
-            if self.local_account_satisfies_clone_request(request_ref) {
-                metrics::inc_chainlink_clone_accounts_total_with_context(
-                    fetch_context.clone(),
-                    remote_result,
-                    clone_intent,
-                    ChainlinkCloneOutcome::Skipped,
-                );
-                return Ok(Signature::default());
-            }
-
-            match self.claim_pending_clone(pubkey) {
-                CloneClaim::Owner => {
-                    let mut guard = PendingCloneGuard::new(
-                        Arc::clone(&self.pending_clones),
-                        pubkey,
-                    );
-                    metrics::inc_chainlink_clone_accounts_total_with_context(
-                        fetch_context.clone(),
-                        remote_result,
-                        clone_intent,
-                        ChainlinkCloneOutcome::Submitted,
-                    );
-                    let Some(owned_request) = request.take() else {
-                        let err = ClonerError::CommittorServiceError(
-                            "owner missing request for clone".to_string(),
-                        );
-                        self.finish_pending_clone(
-                            pubkey,
-                            CloneCompletion::Failed,
-                        );
-                        guard.dismiss();
-                        return Err(ClonerError::FailedToCloneRegularAccount(
-                            pubkey,
-                            Box::new(err),
-                        ));
-                    };
-                    let active_delegation_satisfies_request =
-                        owned_request.delegation_actions.is_empty()
-                            && owned_request.delegated_to_other.is_none()
-                            && !owned_request.needs_undelegation;
-                    let is_empty_placeholder =
-                        Self::is_empty_placeholder_account(
-                            &owned_request.account,
-                        );
-                    let request_slot = owned_request.account.remote_slot();
-                    Self::record_empty_placeholder_stage(
-                        is_empty_placeholder,
-                        fetch_context.clone(),
-                        ChainlinkEmptyPlaceholderStage::CloneSubmitted,
-                        Outcome::Success,
-                    );
-                    let result = self.cloner.clone_account(owned_request).await;
-                    let reconciled_active_delegation = result.is_err()
-                        && active_delegation_satisfies_request
-                        && self.accounts_bank.get_account(&pubkey).is_some_and(
-                            |account| {
-                                Self::account_is_actively_delegated(&account)
-                            },
-                        );
-                    if reconciled_active_delegation {
-                        debug!(
-                            pubkey = %pubkey,
-                            error = ?result,
-                            "Clone request satisfied by concurrently active local delegation"
-                        );
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::Skipped,
-                        );
-                    } else if result.is_ok() {
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::CloneSucceeded,
-                        );
-                        let materialization_outcome = self
-                            .record_account_materialization(
-                                &pubkey,
-                                request_slot,
-                                remote_result,
-                                fetch_context.clone(),
-                            );
-                        Self::record_empty_placeholder_materialization_stage(
-                            is_empty_placeholder,
-                            fetch_context.clone(),
-                            materialization_outcome,
-                        );
-                    } else {
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::CloneFailed,
-                        );
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::SubmitFailed,
-                        );
-                        Self::record_empty_placeholder_stage(
-                            is_empty_placeholder,
-                            fetch_context.clone(),
-                            ChainlinkEmptyPlaceholderStage::CloneSubmitFailed,
-                            Outcome::Error,
-                        );
-                    }
-                    let completion =
-                        if result.is_ok() || reconciled_active_delegation {
-                            CloneCompletion::Success
-                        } else {
-                            CloneCompletion::Failed
-                        };
-                    self.finish_pending_clone(pubkey, completion);
-                    guard.dismiss();
-                    return if reconciled_active_delegation {
-                        Ok(Signature::default())
-                    } else {
-                        result
-                    };
-                }
-                CloneClaim::Waiter(rx) => match rx.await {
-                    Ok(CloneCompletion::Success) => continue,
-                    Ok(CloneCompletion::Failed) => {
-                        return Err(ClonerError::FailedToCloneRegularAccount(
-                            pubkey,
-                            Box::new(ClonerError::CommittorServiceError(
-                                "Clone owner failed".to_string(),
-                            )),
-                        ));
-                    }
-                    Err(_) => {
-                        return Err(ClonerError::FailedToCloneRegularAccount(
-                            pubkey,
-                            Box::new(ClonerError::CommittorServiceError(
-                                "Clone owner dropped".to_string(),
-                            )),
-                        ));
-                    }
-                },
-            }
-        }
-    }
-
-    async fn clone_program_with_ownership(
-        &self,
-        program: LoadedProgram,
-        fetch_context: AccountFetchContext,
-    ) -> ClonerResult<Signature> {
-        let program_id = program.program_id;
-        let remote_slot = program.remote_slot;
-        let remote_result = ChainlinkCloneRemoteResult::Found;
-        let clone_intent = ChainlinkCloneIntent::ProgramData;
-
-        loop {
-            if self
-                .accounts_bank
-                .get_account(&program_id)
-                .is_some_and(|account| account.remote_slot() >= remote_slot)
-            {
-                metrics::inc_chainlink_clone_accounts_total_with_context(
-                    fetch_context.clone(),
-                    remote_result,
-                    clone_intent,
-                    ChainlinkCloneOutcome::Skipped,
-                );
-                return Ok(Signature::default());
-            }
-
-            match self.claim_pending_clone(program_id) {
-                CloneClaim::Owner => {
-                    let mut guard = PendingCloneGuard::new(
-                        Arc::clone(&self.pending_clones),
-                        program_id,
-                    );
-
-                    let result = if self
-                        .accounts_bank
-                        .get_account(&program_id)
-                        .is_some_and(|account| {
-                            account.remote_slot() >= remote_slot
-                        }) {
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::Skipped,
-                        );
-                        Ok(Signature::default())
-                    } else {
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::Submitted,
-                        );
-                        let result = self.cloner.clone_program(program).await;
-                        if result.is_ok() {
-                            metrics::inc_chainlink_clone_accounts_total_with_context(
-                                fetch_context.clone(),
-                                remote_result,
-                                clone_intent,
-                                ChainlinkCloneOutcome::CloneSucceeded,
-                            );
-                            self.record_program_materialization(
-                                &program_id,
-                                remote_slot,
-                                fetch_context,
-                            );
-                        } else {
-                            metrics::inc_chainlink_clone_accounts_total_with_context(
-                                fetch_context.clone(),
-                                remote_result,
-                                clone_intent,
-                                ChainlinkCloneOutcome::CloneFailed,
-                            );
-                            metrics::inc_chainlink_clone_accounts_total_with_context(
-                                fetch_context.clone(),
-                                remote_result,
-                                clone_intent,
-                                ChainlinkCloneOutcome::SubmitFailed,
-                            );
-                        }
-                        result
-                    };
-                    let completion = if result.is_ok() {
-                        CloneCompletion::Success
-                    } else {
-                        CloneCompletion::Failed
-                    };
-                    self.finish_pending_clone(program_id, completion);
-                    guard.dismiss();
-                    return result;
-                }
-                CloneClaim::Waiter(rx) => match rx.await {
-                    Ok(CloneCompletion::Success) => continue,
-                    Ok(CloneCompletion::Failed) => {
-                        return Err(ClonerError::FailedToCloneProgram(
-                            program_id,
-                            Box::new(ClonerError::CommittorServiceError(
-                                "Clone owner failed".to_string(),
-                            )),
-                        ));
-                    }
-                    Err(_) => {
-                        return Err(ClonerError::FailedToCloneProgram(
-                            program_id,
-                            Box::new(ClonerError::CommittorServiceError(
-                                "Clone owner dropped".to_string(),
-                            )),
-                        ));
-                    }
-                },
-            }
-        }
-    }
-
-    async fn clone_account_with_post_delegation_action_invariants(
-        &self,
-        mut request: AccountCloneRequest,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<Signature> {
-        if request.account.delegated()
-            && is_ata(&request.pubkey, &request.account).is_some()
-            && !normalize_native_token_account_for_local_clone(
-                &mut request.account,
-            )
-        {
-            return Err(ChainlinkError::InvalidTokenAccount(
-                request.pubkey,
-                "delegated ATA token data is malformed".to_string(),
-            ));
-        }
-        self.normalize_unresolved_dlp_clone_request(&mut request)?;
-
-        if request.account.delegated()
-            && self.local_delegated_clone_target_active(request.pubkey)
-        {
-            return Ok(Signature::default());
-        }
-
-        if request.delegation_actions.is_empty() {
-            return Ok(self
-                .clone_account_with_ownership(request, fetch_context)
-                .await?);
-        }
-
-        if !request.account.delegated() {
-            return Err(ChainlinkError::InvalidDelegationActions(
-                request.pubkey,
-                "post-delegation actions attached to non-delegated clone target"
-                    .to_string(),
-            ));
-        }
-
-        let result = async {
-            self.ensure_delegation_action_dependencies(
-                request.pubkey,
-                request.account.remote_slot(),
-                &request.delegation_actions,
-                fetch_context.clone(),
-            )
-            .await?;
-
-            Ok(self
-                .clone_account_with_ownership(
-                    request.clone(),
-                    fetch_context.clone(),
-                )
-                .await?)
-        }
-        .await;
-
-        match result {
-            Ok(signature) => Ok(signature),
-            Err(err) => {
-                let pubkey = request.pubkey;
-                if self
-                    .accounts_bank
-                    .get_account(&pubkey)
-                    .is_some_and(|account| account.undelegating())
-                {
-                    return Err(err);
-                }
-
-                let Some(_guard) = self.claim_undelegation(pubkey) else {
-                    return Err(err);
-                };
-
-                // The post-delegation actions could not be satisfied (e.g. a
-                // high-risk signer or a missing dependency): the account is
-                // still cloned, but flagged so it gets automatically
-                // undelegated back to chain.
-                match self
-                    .clone_account_and_schedule_undelegation_with_ownership(
-                        request,
-                        fetch_context.clone(),
-                    )
-                    .await
-                {
-                    Ok(signature) => Ok(signature),
-                    Err(undelegation_err) => {
-                        warn!(
-                            pubkey = %pubkey,
-                            error = ?err,
-                            undelegation_error = ?undelegation_err,
-                            "Failed to schedule undelegation after post-delegation action clone failure"
-                        );
-                        Err(err)
-                    }
-                }
-            }
-        }
-    }
-
-    async fn clone_account_and_schedule_undelegation_with_ownership(
-        &self,
-        mut request: AccountCloneRequest,
-        fetch_context: AccountFetchContext,
-    ) -> ClonerResult<Signature> {
-        let pubkey = request.pubkey;
-        request.needs_undelegation = true;
-        let remote_result = Self::clone_remote_result_for_request(&request);
-        let clone_intent = Self::clone_intent_for_request(&request);
-        let mut request = Some(request);
-
-        loop {
-            if self
-                .accounts_bank
-                .get_account(&pubkey)
-                .is_some_and(|account| account.undelegating())
-            {
-                metrics::inc_chainlink_clone_accounts_total_with_context(
-                    fetch_context.clone(),
-                    remote_result,
-                    clone_intent,
-                    ChainlinkCloneOutcome::Skipped,
-                );
-                return Ok(Signature::default());
-            }
-
-            match self.claim_pending_clone(pubkey) {
-                CloneClaim::Owner => {
-                    let mut guard = PendingCloneGuard::new(
-                        Arc::clone(&self.pending_clones),
-                        pubkey,
-                    );
-                    let Some(owned_request) = request.take() else {
-                        let err = ClonerError::CommittorServiceError(
-                            "owner missing request for undelegation clone"
-                                .to_string(),
-                        );
-                        self.finish_pending_clone(
-                            pubkey,
-                            CloneCompletion::Failed,
-                        );
-                        guard.dismiss();
-                        return Err(
-                            ClonerError::FailedToCloneAndScheduleUndelegation(
-                                pubkey,
-                                Box::new(err),
-                            ),
-                        );
-                    };
-                    metrics::inc_chainlink_clone_accounts_total_with_context(
-                        fetch_context.clone(),
-                        remote_result,
-                        clone_intent,
-                        ChainlinkCloneOutcome::Submitted,
-                    );
-                    let is_empty_placeholder =
-                        Self::is_empty_placeholder_account(
-                            &owned_request.account,
-                        );
-                    let request_slot = owned_request.account.remote_slot();
-                    Self::record_empty_placeholder_stage(
-                        is_empty_placeholder,
-                        fetch_context.clone(),
-                        ChainlinkEmptyPlaceholderStage::CloneSubmitted,
-                        Outcome::Success,
-                    );
-                    let result = self.cloner.clone_account(owned_request).await;
-                    if result.is_ok() {
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::CloneSucceeded,
-                        );
-                        let materialization_outcome = self
-                            .record_account_materialization(
-                                &pubkey,
-                                request_slot,
-                                remote_result,
-                                fetch_context.clone(),
-                            );
-                        Self::record_empty_placeholder_materialization_stage(
-                            is_empty_placeholder,
-                            fetch_context.clone(),
-                            materialization_outcome,
-                        );
-                    } else {
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::CloneFailed,
-                        );
-                        metrics::inc_chainlink_clone_accounts_total_with_context(
-                            fetch_context.clone(),
-                            remote_result,
-                            clone_intent,
-                            ChainlinkCloneOutcome::SubmitFailed,
-                        );
-                        Self::record_empty_placeholder_stage(
-                            is_empty_placeholder,
-                            fetch_context.clone(),
-                            ChainlinkEmptyPlaceholderStage::CloneSubmitFailed,
-                            Outcome::Error,
-                        );
-                    }
-                    let completion = if result.is_ok() {
-                        CloneCompletion::Success
-                    } else {
-                        CloneCompletion::Failed
-                    };
-                    self.finish_pending_clone(pubkey, completion);
-                    guard.dismiss();
-                    return result;
-                }
-                CloneClaim::Waiter(rx) => match rx.await {
-                    Ok(CloneCompletion::Success) => continue,
-                    Ok(CloneCompletion::Failed) => {
-                        return Err(ClonerError::FailedToCloneRegularAccount(
-                            pubkey,
-                            Box::new(ClonerError::CommittorServiceError(
-                                "Clone owner failed".to_string(),
-                            )),
-                        ));
-                    }
-                    Err(_) => {
-                        return Err(ClonerError::FailedToCloneRegularAccount(
-                            pubkey,
-                            Box::new(ClonerError::CommittorServiceError(
-                                "Clone owner dropped".to_string(),
-                            )),
-                        ));
-                    }
-                },
-            }
-        }
-    }
-
-    fn claim_undelegation(
-        &self,
-        pubkey: Pubkey,
-    ) -> Option<PendingUndelegationGuard> {
-        let mut pending_undelegations =
-            self.pending_undelegations.lock().ok()?;
-        if !pending_undelegations.insert(pubkey) {
-            return None;
-        }
-        Some(PendingUndelegationGuard {
-            pending_undelegations: Arc::clone(&self.pending_undelegations),
-            pubkey,
-        })
-    }
-
-    fn normalize_unresolved_dlp_clone_request(
-        &self,
-        request: &mut AccountCloneRequest,
-    ) -> ChainlinkResult<()> {
-        if request.account.owner() != &dlp_api::id()
-            || !request.account.delegated()
-        {
-            return Ok(());
-        }
-
-        if request.pubkey
-            == dlp_api::pda::magic_fee_vault_pda_from_validator(
-                &self.validator_pubkey,
-            )
-        {
-            return Ok(());
-        }
-
-        if !request.delegation_actions.is_empty() {
-            return Err(ChainlinkError::InvalidDelegationActions(
-                request.pubkey,
-                "post-delegation actions attached to unresolved DLP-owned clone target"
-                    .to_string(),
-            ));
-        }
-
-        request.account.set_delegated(false);
-        request.account.set_confined(false);
-        Ok(())
-    }
-
-    fn local_delegated_clone_target_active(&self, pubkey: Pubkey) -> bool {
-        self.accounts_bank
-            .get_account(&pubkey)
-            .is_some_and(|account| account.delegated())
-    }
-
-    pub fn start_subscription_listener(
-        self: Arc<Self>,
-        mut subscription_updates: mpsc::Receiver<ForwardedSubscriptionUpdate>,
-    ) {
-        tokio::spawn(async move {
-            let semaphore =
-                Arc::new(Semaphore::new(super::SUBSCRIPTION_UPDATE_LIMIT));
-            let mut pending_tasks: JoinSet<()> = JoinSet::new();
-
-            loop {
-                while let Some(result) = pending_tasks.try_join_next() {
-                    if let Err(err) = result {
-                        warn!(error = ?err, "Subscription update task panicked");
-                    }
-                }
-
-                // INVARIANT: The semaphore is created locally and never closed,
-                // so acquire_owned() cannot fail with AcquireError.
-                let permit = Arc::clone(&semaphore)
-                    .acquire_owned()
-                    .await
-                    .expect("subscription update semaphore never closed");
-
-                match subscription_updates.recv().await {
-                    Some(update) => {
-                        let pubkey = update.pubkey;
-                        trace!(
-                            pubkey = %pubkey,
-                            "FetchCloner received subscription update"
-                        );
-                        let this = Arc::clone(&self);
-                        metrics::inc_inflight_subscription_updates();
-                        pending_tasks.spawn(async move {
-                            struct InflightSubscriptionUpdateGuard;
-                            impl Drop for InflightSubscriptionUpdateGuard {
-                                fn drop(&mut self) {
-                                    metrics::dec_inflight_subscription_updates(
-                                    );
-                                }
-                            }
-                            let _inflight_guard =
-                                InflightSubscriptionUpdateGuard;
-
-                            Self::process_subscription_update(
-                                &this, pubkey, update,
-                            )
-                            .await;
-                            drop(permit);
-                        });
-                    }
-                    None => {
-                        drop(permit);
-                        while pending_tasks.join_next().await.is_some() {}
-                        break;
-                    }
-                }
-            }
-        });
-    }
-
-    async fn process_subscription_update(
-        &self,
-        pubkey: Pubkey,
-        update: ForwardedSubscriptionUpdate,
-    ) {
-        let fresh_update_account = update.account.fresh_account();
-        let is_dlp_owned_update = fresh_update_account
-            .as_ref()
-            .is_some_and(|account| account.owner() == &dlp_api::id());
-        let is_internal_dlp_update =
-            fresh_update_account.as_ref().is_some_and(|account| {
-                is_internal_dlp_account_data(account.data())
-            });
-
-        let dlp_program_interest =
-            if matches!(update.source, SubscriptionSource::Program)
-                && is_dlp_owned_update
-            {
-                match fresh_update_account.as_ref() {
-                    Some(account) => Some(
-                        self.classify_dlp_program_update_interest(
-                            pubkey, account,
-                        )
-                        .await,
-                    ),
-                    None => None,
-                }
-            } else {
-                None
-            };
-
-        match dlp_program_interest {
-            Some(DlpProgramUpdateInterest::DropLocalDelegatedAuthoritative) => {
-                self.cleanup_direct_subscription_for_delegated_account(pubkey)
-                    .await;
-                trace!(
-                    pubkey = %pubkey,
-                    "Dropping DLP program update for locally authoritative delegated account"
-                );
-                return;
-            }
-            Some(DlpProgramUpdateInterest::ProcessUndelegating)
-            | Some(DlpProgramUpdateInterest::ProcessAtaProjection)
-            | Some(DlpProgramUpdateInterest::ProcessDirectlyWatched)
-            | Some(DlpProgramUpdateInterest::DiscoverDelegatedAccount)
-            | None => {}
-        }
-
-        // Internal DLP payloads (records/metadata/commit state) can never be
-        // greedily cloned, so drop them before discovery issues remote
-        // fetches. The exception is an account whose app data collides with
-        // an internal discriminator: its delegation also writes the
-        // delegation record, which the mirror observes on its own stream —
-        // resolve via the mirror now, or after one short recheck covering
-        // the skew between the account and record streams. Only the program
-        // firehose is handled here; account-sourced internal payloads (e.g.
-        // directly watched records) proceed to normal processing.
-        if !matches!(
-            dlp_program_interest,
-            Some(DlpProgramUpdateInterest::ProcessUndelegating)
-                | Some(DlpProgramUpdateInterest::ProcessAtaProjection)
-        ) && is_dlp_owned_update
-            && is_internal_dlp_update
-            && matches!(update.source, SubscriptionSource::Program)
-        {
-            // Internal-shaped bytes cannot be told apart from a colliding
-            // user account (record-shaped app data included), so every
-            // payload is probed; genuine internal accounts have no record
-            // at their derived PDA and drop after the recheck.
-            self.resolve_internal_dlp_collision(pubkey, update.account.slot())
-                .await;
-            return;
-        }
-
-        if self
-            .maybe_greedily_clone_discovered_delegated_account(pubkey, &update)
-            .await
-        {
-            return;
-        }
-        // A late forwarded update can arrive after an account was removed from
-        // the provider watch set. If a new subscription already won the race,
-        // is_watching is true and this update can be processed normally. If this
-        // update wins before acquire_subscription completes, the update is dropped;
-        // the new subscription path performs its own fetch and clones fresh state.
-        // If stale state is still present locally, cleanup is routed through the
-        // existing removal listener, which serializes the final is_watching check and
-        // eviction submission against same-pubkey subscription transitions.
-        //
-        // The guard only applies to account-subscription updates: the
-        // account-sub LRU is the source of truth for `is_watching`. Program
-        // subscription updates can legitimately arrive for pubkeys that are
-        // *not* in the account-sub LRU (e.g. delegated accounts whose direct
-        // subscription was released after cloning and are now tracked only via
-        // their owner program). Dropping those would leave the bank stuck in a
-        // stale delegated/undelegated state.
-        let update_slot = update.account.slot();
-        if matches!(update.source, SubscriptionSource::Account)
-            && !self.remote_account_provider.is_watching(&pubkey)
-        {
-            trace!(
-                pubkey = %pubkey,
-                update_slot,
-                "Dropping subscription update for account that is no longer watched"
-            );
-            if self.accounts_bank.get_account(&pubkey).is_some() {
-                if let Err(err) = self
-                    .remote_account_provider
-                    .send_removal_update(pubkey)
-                    .await
-                {
-                    warn!(
-                        pubkey = %pubkey,
-                        error = ?err,
-                        "Failed to enqueue stale subscription update removal"
-                    );
-                }
-            }
-            return;
-        }
-
-        let companion_fetch_log_context = CompanionFetchLogContext {
-            origin: AccountFetchContext::subscription_update(
-                AccountFetchReason::SubscriptionUpdateClone,
-            ),
-            primary_pubkey: pubkey,
-            context_slot: update_slot,
-        };
-
-        let update_source = update.source;
-        let (resolved_account, deleg_record, delegation_actions) = self
-            .resolve_account_to_clone_from_forwarded_sub_with_unsubscribe(
-                update,
-                &companion_fetch_log_context,
-            )
-            .await;
-        let Some(account) = resolved_account else {
-            return;
-        };
-        let subscription_clone_context =
-            AccountFetchContext::subscription_update(
-                AccountFetchReason::SubscriptionUpdateClone,
-            );
-        let projected_ata_clone_request = self
-            .maybe_build_projected_ata_clone_request_from_subscription_update_with_source(
-                pubkey,
-                &account,
-                update_source,
-                deleg_record.as_ref(),
-                &delegation_actions,
-                &companion_fetch_log_context,
-            )
-            .await;
-
-        //
-        // Ensure that the subscription update isn't out of order, i.e.
-        // we already hold a newer version of the account in our bank.
-        //
-        // The stricter intent is to ignore non-advancing subscription updates: if the bank
-        // already has the account at the same slot, then a normal/plain update at that slot is
-        // treated as stale/duplicate and should not overwrite local state, with the following
-        // exception:
-        //
-        //  - In the undelegate/redelegate same-slot path, the bank can still hold a plain
-        //    or undelegating version while the subscription update carries the delegated state
-        //    at the same slot, so we must allow that update.
-        //
-        let non_advancing_slot =
-            self.accounts_bank.get_account(&pubkey).and_then(|in_bank| {
-                let bank_slot = in_bank.remote_slot();
-                let update_slot = account.remote_slot();
-                let same_slot_delegated_refresh = bank_slot == update_slot
-                    && account.delegated()
-                    && (!in_bank.delegated() || in_bank.undelegating());
-                if bank_slot > update_slot
-                    || (bank_slot == update_slot
-                        && !same_slot_delegated_refresh)
-                {
-                    Some(bank_slot)
-                } else {
-                    None
-                }
-            });
-
-        if let Some(in_bank_slot) = non_advancing_slot {
-            let update_slot = account.remote_slot();
-            if in_bank_slot == update_slot {
-                if let Some(projected_ata_clone_request) =
-                    projected_ata_clone_request
-                {
-                    if let Err(err) = self
-                        .clone_projected_ata_request(
-                            projected_ata_clone_request,
-                            subscription_clone_context,
-                        )
-                        .await
-                    {
-                        warn!(
-                            pubkey = %pubkey,
-                            error = %err,
-                            "Failed to clone projected ATA from out-of-order delegated eATA update"
-                        );
-                    }
-                }
-            }
-            trace!(
-                pubkey = %pubkey,
-                bank_slot = in_bank_slot,
-                update_slot,
-                "Ignoring out-of-order subscription update"
-            );
-            return;
-        }
-
-        let mut undelegation_completed_on_chain = false;
-        if let Some(in_bank) = self.accounts_bank.get_account(&pubkey) {
-            if in_bank.delegated() && !in_bank.undelegating() {
-                self.cleanup_direct_subscription_for_delegated_account(pubkey)
-                    .await;
-                return;
-            }
-
-            if in_bank.undelegating() {
-                debug!(
-                    pubkey = %pubkey,
-                    in_bank_delegated = in_bank.delegated(),
-                    in_bank_owner = %in_bank.owner(),
-                    in_bank_slot = in_bank.remote_slot(),
-                    chain_delegated = account.delegated(),
-                    chain_owner = %account.owner(),
-                    chain_slot = account.remote_slot(),
-                    "Received update for undelegating account"
-                );
-
-                if account.delegated()
-                    && ata_projection::derive_eata_pubkey_from_ata_account(
-                        &pubkey, &account,
-                    )
-                    .is_some()
-                    && deleg_record.as_ref().is_some_and(|record| {
-                        record.owner == EATA_PROGRAM_ID
-                            && record.authority == self.validator_pubkey
-                    })
-                {
-                    debug!(
-                        pubkey = %pubkey,
-                        "Keeping undelegating ATA in bank while companion eATA remains delegated"
-                    );
-                    return;
-                }
-
-                // This will only be true in the following case:
-                // 1. a commit was triggered for the account
-                // 2. a commit + undelegate was triggered for the account -> undelegating
-                // 3. we receive the update for (1.)
-                //
-                // Thus our state is more up to date and we don't
-                // need to update our bank.
-                if account_still_undelegating_on_chain(
-                    &pubkey,
-                    account.delegated(),
-                    in_bank.remote_slot(),
-                    deleg_record,
-                    &self.validator_pubkey,
-                ) {
-                    return;
-                }
-                undelegation_completed_on_chain = true;
-            } else if !in_bank.delegated() && account.delegated() {
-                undelegation_completed_on_chain = true;
-            } else if in_bank.owner().eq(&dlp_api::id()) {
-                debug!(
-                    pubkey = %pubkey,
-                    "Received update for account owned by delegation program but not marked as undelegating"
-                );
-            }
-        } else {
-            debug!(
-                pubkey = %pubkey,
-                "Received update for account not in bank"
-            );
-            if account.delegated() {
-                undelegation_completed_on_chain = true;
-            }
-        }
-
-        // Determine if delegated to another validator
-        let delegated_to_other = deleg_record
-            .as_ref()
-            .and_then(|dr| self.get_delegated_to_other(dr));
-
-        // Delegated subscription cleanup is limited to direct subscription/LRU
-        // ownership here; undelegation tracking owns protected subscriptions
-        // until undelegation is explicitly complete.
-        if undelegation_completed_on_chain {
-            if !account.delegated() {
-                self.ensure_direct_subscription_for_completed_account(pubkey)
-                    .await;
-            }
-            self.cleanup_undelegation_tracking_for_completed_account(pubkey)
-                .await;
-        }
-        if account.delegated() {
-            self.cleanup_direct_subscription_for_delegated_account(pubkey)
-                .await;
-        }
-
-        if account.executable() {
-            self.handle_executable_sub_update(
-                pubkey,
-                account,
-                &companion_fetch_log_context,
-            )
-            .await;
-        } else {
-            let commit_frequency_ms = deleg_record.as_ref().and_then(|dr| {
-                dr.authority
-                    .eq(&self.validator_pubkey)
-                    .then_some(dr.commit_frequency_ms)
-            });
-            let raw_delegation_actions = if account.delegated()
-                && projected_ata_clone_request.is_none()
-            {
-                delegation_actions
-            } else {
-                DelegationActions::default()
-            };
-            if let Err(err) = self
-                .clone_account_with_post_delegation_action_invariants(
-                    AccountCloneRequest {
-                        pubkey,
-                        account,
-                        commit_frequency_ms,
-                        delegation_actions: raw_delegation_actions,
-                        delegated_to_other,
-                        needs_undelegation: false,
-                    },
-                    subscription_clone_context.clone(),
-                )
-                .await
-            {
-                error!(
-                    pubkey = %pubkey,
-                    error = %err,
-                    "Failed to clone account into bank"
-                );
-            } else if let Some(projected_ata_clone_request) =
-                projected_ata_clone_request
-            {
-                if let Err(err) = self
-                    .clone_projected_ata_request(
-                        projected_ata_clone_request,
-                        subscription_clone_context.clone(),
-                    )
-                    .await
-                {
-                    error!(
-                        pubkey = %pubkey,
-                        error = %err,
-                        "Failed to clone projected ATA from delegated eATA update"
-                    );
-                }
-            }
-        }
-    }
-
-    async fn ensure_delegation_action_dependencies(
-        &self,
-        pubkey: Pubkey,
-        remote_slot: u64,
-        delegation_actions: &DelegationActions,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<()> {
-        if delegation_actions.is_empty() {
-            return Ok(());
-        }
-
-        self.validate_post_delegation_action_signers(delegation_actions)
-            .await?;
-
-        let (dependencies, writable_dependencies) =
-            Self::collect_post_delegation_action_dependencies(
-                pubkey,
-                delegation_actions,
-            );
-
-        let dependencies_to_fetch = dependencies
-            .into_iter()
-            .filter(|dependency| {
-                self.delegation_action_dependency_needs_fetch(
-                    dependency,
-                    &writable_dependencies,
-                )
-            })
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect::<Vec<_>>();
-
-        if dependencies_to_fetch.is_empty() {
-            return Ok(());
-        }
-
-        let result = self
-            .fetch_and_clone_accounts_with_dedup_forced_refresh(
-                &dependencies_to_fetch,
-                None,
-                Some(remote_slot),
-                fetch_context.with_reason(
-                    AccountFetchReason::ActionDependencyForcedRefresh,
-                ),
-                &writable_dependencies,
-                None,
-            )
-            .await?;
-        if result.missing_delegation_record.is_empty() {
-            return Ok(());
-        }
-
-        let missing_accounts = result
-            .pubkeys_missing_delegation_record()
-            .into_iter()
-            .collect::<HashSet<_>>()
-            .into_iter()
-            .collect::<Vec<_>>();
-        let mut missing_accounts = missing_accounts;
-        missing_accounts.sort_unstable();
-        Err(ChainlinkError::MissingDelegationActionAccounts(
-            missing_accounts,
-        ))
-    }
-
-    fn collect_post_delegation_action_dependencies(
-        target: Pubkey,
-        delegation_actions: &DelegationActions,
-    ) -> (HashSet<Pubkey>, HashSet<Pubkey>) {
-        let mut dependencies = HashSet::new();
-        let mut writable_dependencies = HashSet::new();
-        for instruction in delegation_actions.iter() {
-            if instruction.program_id != target {
-                dependencies.insert(instruction.program_id);
-            }
-            for meta in &instruction.accounts {
-                if meta.pubkey == target {
-                    continue;
-                }
-                dependencies.insert(meta.pubkey);
-                if meta.is_writable {
-                    writable_dependencies.insert(meta.pubkey);
-                }
-            }
-        }
-        (dependencies, writable_dependencies)
-    }
-
-    fn delegation_action_dependency_needs_fetch(
-        &self,
-        dependency: &Pubkey,
-        writable_dependencies: &HashSet<Pubkey>,
-    ) -> bool {
-        let Some(account) = self.accounts_bank.get_account(dependency) else {
-            return true;
-        };
-        writable_dependencies.contains(dependency)
-            && (!account.delegated() || account.undelegating())
-    }
-
-    async fn validate_post_delegation_action_signers(
-        &self,
-        delegation_actions: &DelegationActions,
-    ) -> ChainlinkResult<()> {
-        let Some(risk_service) = self.risk_service.as_ref() else {
-            return Ok(());
-        };
-
-        let mut signers = delegation_actions
-            .iter()
-            .flat_map(|instruction| {
-                instruction.accounts.iter().filter_map(|meta| {
-                    if meta.is_signer {
-                        Some(meta.pubkey.to_string())
-                    } else {
-                        None
-                    }
-                })
-            })
-            .collect::<Vec<_>>();
-        signers.sort_unstable();
-        signers.dedup();
-
-        if signers.is_empty() {
-            return Ok(());
-        }
-        Ok(risk_service.check_addresses(signers).await?)
-    }
-
-    async fn clone_projected_ata_request(
-        &self,
-        request: AccountCloneRequest,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<Signature> {
-        if self
-            .accounts_bank
-            .get_account(&request.pubkey)
-            .is_some_and(|account| account.undelegating())
-        {
-            return Ok(Signature::default());
-        }
-
-        self.clone_account_with_post_delegation_action_invariants(
-            request,
-            fetch_context.with_reason(AccountFetchReason::AtaProjection),
-        )
-        .await
-    }
-
-    /// A program-source internal-looking update is either genuine internal
-    /// DLP state or a delegated account whose app data collides with an
-    /// internal discriminator. Its delegation writes the record in the same
-    /// slot, so the mirror proves the collision: resolve immediately on a
-    /// mirror hit, or once more after a short delay covering record-stream
-    /// skew. A final miss drops the update — the on-demand ensure path
-    /// remains the backstop, exactly as unsighted parked candidates were
-    /// dropped before. Without a live mirror there is no proactive
-    /// collision discovery.
-    async fn resolve_internal_dlp_collision(&self, pubkey: Pubkey, slot: u64) {
-        let Some(mirror) = &self.record_mirror else {
-            trace!(
-                pubkey = %pubkey,
-                "Dropping internal DLP program subscription update (no record mirror)"
-            );
-            return;
-        };
-        let record_pda = delegation_record_pda_from_delegated_account(&pubkey);
-        if matches!(mirror.probe(&record_pda, slot), MirrorLookup::Hit { .. }) {
-            self.clone_colliding_delegated_account(pubkey, slot).await;
-        } else {
-            self.schedule_collision_recheck(pubkey, slot);
-        }
-    }
-
-    /// Parks a collision candidate for delayed mirror rechecks. Dedup is
-    /// keyed by the derived record PDA with a monotonic slot. Admitted
-    /// candidates are never displaced: at capacity, newcomers are dropped
-    /// instead — parity with the parking this replaces, so genuine internal
-    /// firehose churn cannot evict a real collision candidate. Exactly one
-    /// worker owns each admitted entry, and only that worker removes it, so
-    /// entries and worker tasks stay 1:1 bounded by the map capacity.
-    fn schedule_collision_recheck(&self, pubkey: Pubkey, slot: u64) {
-        let record_pda = delegation_record_pda_from_delegated_account(&pubkey);
-        {
-            let mut rechecks = self.pending_collision_rechecks.lock();
-            if let Some(pending) = rechecks.get_mut(&record_pda) {
-                // A replayed older update must not downgrade the slot.
-                pending.1 = pending.1.max(slot);
-                return;
-            }
-            if rechecks.len() >= PENDING_COLLISION_RECHECKS_CAPACITY.get() {
-                trace!(
-                    pubkey = %pubkey,
-                    slot,
-                    "Dropping collision candidate; recheck queue is full"
-                );
-                return;
-            }
-            rechecks.put(record_pda, (pubkey, slot));
-        }
-
-        let this = self.clone();
-        task::spawn(async move {
-            // Probes for the current generation, i.e. consecutive attempts
-            // that observed the same slot. A newer update bumps the slot in
-            // place and restarts the budget; every bump is driven by a real
-            // firehose update, which rate-limits the loop externally.
-            let mut generation_probes = 0usize;
-            let mut probed_slot = None;
-            loop {
-                let delay_idx =
-                    generation_probes.min(COLLISION_RECHECK_DELAYS.len() - 1);
-                tokio::time::sleep(COLLISION_RECHECK_DELAYS[delay_idx]).await;
-
-                // Read fresh each attempt: later updates bump the slot.
-                let candidate = this
-                    .pending_collision_rechecks
-                    .lock()
-                    .peek(&record_pda)
-                    .copied();
-                let Some((pubkey, slot)) = candidate else {
-                    return;
-                };
-
-                let hit = this.record_mirror.as_ref().is_some_and(|mirror| {
-                    matches!(
-                        mirror.probe(&record_pda, slot),
-                        MirrorLookup::Hit { .. }
-                    )
-                });
-                if hit {
-                    // Remove only the generation that was probed: a newer
-                    // update racing in keeps the entry for the next attempt.
-                    let owned = {
-                        let mut rechecks =
-                            this.pending_collision_rechecks.lock();
-                        let matches_probed = rechecks
-                            .peek(&record_pda)
-                            .is_some_and(|&(_, s)| s == slot);
-                        if matches_probed {
-                            rechecks.pop(&record_pda);
-                        }
-                        matches_probed
-                    };
-                    if owned {
-                        this.clone_colliding_delegated_account(pubkey, slot)
-                            .await;
-                        return;
-                    }
-                    generation_probes = 0;
-                    probed_slot = None;
-                    continue;
-                }
-
-                if probed_slot == Some(slot) {
-                    generation_probes += 1;
-                } else {
-                    generation_probes = 1;
-                    probed_slot = Some(slot);
-                }
-                if generation_probes >= COLLISION_RECHECK_DELAYS.len() {
-                    // Give up only on the generation that exhausted its
-                    // probes; a newer one racing in keeps the worker going.
-                    let dropped = {
-                        let mut rechecks =
-                            this.pending_collision_rechecks.lock();
-                        let matches_probed = rechecks
-                            .peek(&record_pda)
-                            .is_some_and(|&(_, s)| s == slot);
-                        if matches_probed {
-                            rechecks.pop(&record_pda);
-                        }
-                        matches_probed
-                    };
-                    if dropped {
-                        trace!(
-                            pubkey = %pubkey,
-                            slot,
-                            "Dropping internal DLP program subscription update (no record after rechecks)"
-                        );
-                        return;
-                    }
-                    generation_probes = 0;
-                    probed_slot = None;
-                }
-            }
-        });
-    }
-
-    /// Discovery for an account whose app data collides with an internal
-    /// DLP discriminator, proven delegated by its mirrored record:
-    /// authority-gated like discovery, then cloned through the deduped
-    /// on-demand path with a forced refresh.
-    async fn clone_colliding_delegated_account(
-        &self,
-        pubkey: Pubkey,
-        slot: u64,
-    ) {
-        // A pre-delegation bank copy must be force-refreshed; only a
-        // delegated copy at the update slot or newer settles the candidate.
-        let fresh_delegated_in_bank = self
-            .accounts_bank
-            .get_account(&pubkey)
-            .is_some_and(|in_bank| {
-                in_bank.delegated() && in_bank.remote_slot() >= slot
-            });
-        if fresh_delegated_in_bank {
-            return;
-        }
-        let fetch_context = AccountFetchContext::subscription_update(
-            AccountFetchReason::SubscriptionUpdateGreedyDiscovery,
-        );
-        // Same authority gate as discovery: accounts delegated to another
-        // validator must not be cloned from the firehose.
-        let record_context = fetch_context
-            .clone()
-            .with_reason(AccountFetchReason::DelegationRecord);
-        let companion_fetch_log_context = CompanionFetchLogContext {
-            origin: record_context.clone(),
-            primary_pubkey: pubkey,
-            context_slot: slot,
-        };
-        let Some((deleg_record, _)) = self
-            .fetch_and_parse_delegation_record(
-                pubkey,
-                slot,
-                record_context,
-                companion_fetch_log_context,
-            )
-            .await
-        else {
-            trace!(
-                pubkey = %pubkey,
-                slot = slot,
-                "Colliding delegated account has no resolvable delegation record; falling back to on-demand cloning"
-            );
-            return;
-        };
-        let is_delegated_to_us = deleg_record.authority
-            == self.validator_pubkey
-            || deleg_record.authority == Pubkey::default();
-        if !is_delegated_to_us {
-            metrics::inc_discovered_dlp_update_delegated_elsewhere();
-            trace!(
-                pubkey = %pubkey,
-                authority = %deleg_record.authority,
-                "Ignoring colliding delegated account delegated elsewhere"
-            );
-            return;
-        }
-        // The deduped fetch can join an in-flight pre-delegation owner and
-        // settle stale; retry until the bank reflects the mirrored delegation
-        // (at the update slot itself only a delegated copy settles).
-        const MAX_RELEASE_CLONE_ATTEMPTS: usize = 3;
-        for _ in 0..MAX_RELEASE_CLONE_ATTEMPTS {
-            // The update may predate a local undelegation that started in
-            // the meantime. Do not let the mirrored record bypass the normal
-            // confirmation check and overwrite protected local state.
-            if self
-                .accounts_bank
-                .get_account(&pubkey)
-                .is_some_and(|in_bank| {
-                    in_bank.undelegating()
-                        && account_still_undelegating_on_chain(
-                            &pubkey,
-                            true,
-                            in_bank.remote_slot(),
-                            Some(deleg_record),
-                            &self.validator_pubkey,
-                        )
-                })
-            {
-                trace!(
-                    pubkey = %pubkey,
-                    slot = slot,
-                    "Ignoring colliding delegated account while local undelegation remains pending"
-                );
-                return;
-            }
-
-            let result = match self
-                .fetch_and_clone_accounts_with_dedup_forced_refresh(
-                    &[pubkey],
-                    None,
-                    Some(slot),
-                    fetch_context.clone(),
-                    &HashSet::from([pubkey]),
-                    Some((pubkey, deleg_record.delegation_slot)),
-                )
-                .await
-            {
-                Ok(result) => result,
-                Err(err) => {
-                    warn!(
-                        pubkey = %pubkey,
-                        error = %err,
-                        "Failed to clone colliding delegated account"
-                    );
-                    return;
-                }
-            };
-            let unresolvable = result
-                .not_found_on_chain
-                .iter()
-                .chain(result.missing_delegation_record.iter())
-                .any(|(missing_pubkey, _)| missing_pubkey == &pubkey);
-            if unresolvable {
-                trace!(
-                    pubkey = %pubkey,
-                    slot = slot,
-                    ?result,
-                    "Colliding delegated account no longer resolvable on chain"
-                );
-                return;
-            }
-            let in_bank = self.accounts_bank.get_account(&pubkey);
-            let settled = in_bank.as_ref().is_some_and(|in_bank| {
-                in_bank.remote_slot() > slot
-                    || (in_bank.remote_slot() == slot && in_bank.delegated())
-            });
-            if settled {
-                return;
-            }
-            if in_bank.is_some_and(|in_bank| {
-                in_bank.undelegating()
-                    && deleg_record.delegation_slot <= in_bank.remote_slot()
-            }) {
-                break;
-            }
-        }
-        // The RPC kept serving pre-delegation (or still-undelegating) state.
-        // Requeue for strictly newer evidence — a fresh record update or the
-        // mirror watermark passing the failed generation — so a later record
-        // state re-triggers discovery instead of waiting for on-demand
-        // access, as the replaced re-parking did.
-        warn!(
-            pubkey = %pubkey,
-            slot,
-            "Colliding delegated account did not settle at the update slot; requeueing for newer record evidence"
-        );
-        self.schedule_collision_recheck(pubkey, slot + 1);
-    }
-
-    async fn maybe_greedily_clone_discovered_delegated_account(
-        &self,
-        pubkey: Pubkey,
-        update: &ForwardedSubscriptionUpdate,
-    ) -> bool {
-        if self.accounts_bank.get_account(&pubkey).is_some() {
-            return false;
-        }
-
-        let Some(account) = update.account.fresh_account() else {
-            return false;
-        };
-
-        if !account.owner().eq(&dlp_api::id()) {
-            return false;
-        }
-
-        let discovery_context = AccountFetchContext::subscription_update(
-            AccountFetchReason::SubscriptionUpdateGreedyDiscovery,
-        );
-        let record_context = discovery_context
-            .clone()
-            .with_reason(AccountFetchReason::DelegationRecord);
-        let companion_fetch_log_context = CompanionFetchLogContext {
-            origin: record_context.clone(),
-            primary_pubkey: pubkey,
-            context_slot: account.remote_slot(),
-        };
-
-        let Some((deleg_record, delegation_actions)) = self
-            .fetch_and_parse_delegation_record(
-                pubkey,
-                account.remote_slot(),
-                record_context,
-                companion_fetch_log_context,
-            )
-            .await
-        else {
-            trace!(
-                pubkey = %pubkey,
-                slot = account.remote_slot(),
-                "Greedy discovery could not resolve delegation record; falling back"
-            );
-            return false;
-        };
-
-        let is_delegated_to_us = deleg_record.authority
-            == self.validator_pubkey
-            || deleg_record.authority == Pubkey::default();
-        if !is_delegated_to_us {
-            metrics::inc_discovered_dlp_update_delegated_elsewhere();
-            trace!(
-                pubkey = %pubkey,
-                authority = %deleg_record.authority,
-                "Ignoring discovered DLP-owned update delegated elsewhere"
-            );
-            return true;
-        }
-        let delegation_actions = delegation_actions.unwrap_or_default();
-
-        let greedy_ata_pubkeys = delegation::parse_raw_eata_pda(
-            &pubkey,
-            account.data(),
-            deleg_record.owner,
-        )
-        .map(|(wallet_owner, mint)| {
-            ata_projection::derive_supported_ata_pubkeys(&wallet_owner, &mint)
-        })
-        .unwrap_or_default();
-        let mut pubkeys_to_clone =
-            Vec::with_capacity(1 + greedy_ata_pubkeys.len());
-        pubkeys_to_clone.push(pubkey);
-        pubkeys_to_clone.extend(greedy_ata_pubkeys.iter().copied().filter(
-            |ata_pubkey| self.accounts_bank.get_account(ata_pubkey).is_none(),
-        ));
-
-        // Keep eATA discovery with its candidate base ATAs in one clone batch
-        // so the normal ATA projection path runs for the same update.
-        let clone_result = if greedy_ata_pubkeys.is_empty() {
-            self.fetch_and_clone_accounts_with_dedup(
-                &pubkeys_to_clone,
-                None,
-                Some(account.remote_slot()),
-                discovery_context.clone(),
-            )
-            .await
-        } else {
-            self.fetch_and_clone_accounts(
-                &pubkeys_to_clone,
-                None,
-                Some(account.remote_slot()),
-                discovery_context.clone(),
-            )
-            .await
-        };
-
-        match clone_result {
-            Ok(result)
-                if result
-                    .not_found_on_chain
-                    .iter()
-                    .all(|(missing_pubkey, _)| missing_pubkey != &pubkey)
-                    && result.missing_delegation_record.iter().all(
-                        |(missing_pubkey, _)| missing_pubkey != &pubkey,
-                    ) =>
-            {
-                let bank_slot = self
-                    .accounts_bank
-                    .get_account(&pubkey)
-                    .map(|in_bank| in_bank.remote_slot());
-                if bank_slot.is_none_or(|slot| slot < account.remote_slot()) {
-                    trace!(
-                        pubkey = %pubkey,
-                        bank_slot,
-                        update_slot = account.remote_slot(),
-                        ?result,
-                        "Greedy clone did not materialize a fresh enough account; falling back"
-                    );
-                    false
-                } else if let Some(projected_ata_clone_request) = self
-                    .maybe_build_projected_ata_clone_request_from_subscription_update_with_source(
-                        pubkey,
-                        &account,
-                        update.source,
-                        Some(&deleg_record),
-                        &delegation_actions,
-                        &CompanionFetchLogContext {
-                            origin: discovery_context.clone(),
-                            primary_pubkey: pubkey,
-                            context_slot: account.remote_slot(),
-                        },
-                    )
-                    .await
-                {
-                    let projected_ata_pubkey =
-                        projected_ata_clone_request.pubkey;
-                    if let Err(err) = self
-                        .clone_projected_ata_request(
-                            projected_ata_clone_request,
-                            discovery_context.clone(),
-                        )
-                        .await
-                    {
-                        warn!(
-                            pubkey = %pubkey,
-                            error = %err,
-                            "Failed to clone projected ATA from greedily discovered delegated eATA"
-                        );
-                        false
-                    } else {
-                        trace!(
-                            pubkey = %pubkey,
-                            ata_pubkey = %projected_ata_pubkey,
-                            slot = account.remote_slot(),
-                            "Greedily cloned delegated account"
-                        );
-                        true
-                    }
-                } else {
-                    let cloned_ata_pubkey = greedy_ata_pubkeys
-                        .iter()
-                        .copied()
-                        .find(|ata_pubkey| {
-                            self.accounts_bank
-                                .get_account(ata_pubkey)
-                                .is_some_and(|account_in_bank| {
-                                    account_in_bank.remote_slot()
-                                        >= account.remote_slot()
-                                })
-                        });
-                    if let Some(ata_pubkey) = cloned_ata_pubkey {
-                        trace!(
-                            pubkey = %pubkey,
-                            ata_pubkey = %ata_pubkey,
-                            slot = account.remote_slot(),
-                            "Greedily cloned delegated account"
-                        );
-                    } else {
-                        trace!(
-                            pubkey = %pubkey,
-                            slot = account.remote_slot(),
-                            "Greedily cloned delegated account"
-                        );
-                    }
-                    true
-                }
-            }
-            Ok(result) => {
-                trace!(
-                    pubkey = %pubkey,
-                    ?result,
-                    "Greedy clone incomplete; falling back"
-                );
-                false
-            }
-            Err(err) => {
-                warn!(
-                    pubkey = %pubkey,
-                    error = %err,
-                    "Failed to greedily clone discovered delegated account"
-                );
-                false
-            }
-        }
-    }
-
-    async fn handle_executable_sub_update(
-        &self,
-        pubkey: Pubkey,
-        account: AccountSharedData,
-        companion_fetch_log_context: &CompanionFetchLogContext,
-    ) {
-        // moved to program_loader module
-        program_loader::handle_executable_sub_update_with_context(
-            self,
-            pubkey,
-            account,
-            companion_fetch_log_context,
-        )
-        .await;
-    }
-
-    async fn cleanup_direct_subscription_for_delegated_account(
-        &self,
-        pubkey: Pubkey,
-    ) {
-        if let Err(err) = self
-            .remote_account_provider
-            .release_subscription_reason_silently_for_delegated_account(
-                &pubkey,
-                SubscriptionReason::DirectAccount,
-            )
-            .await
-        {
-            warn!(
-                pubkey = %pubkey,
-                error = %err,
-                "Failed to clean up direct subscription for delegated account"
-            );
-        }
-    }
-
-    async fn ensure_direct_subscription_for_completed_account(
-        &self,
-        pubkey: Pubkey,
-    ) {
-        if let Err(err) = self
-            .remote_account_provider
-            .ensure_subscription(&pubkey, SubscriptionReason::DirectAccount)
-            .await
-        {
-            warn!(
-                pubkey = %pubkey,
-                error = %err,
-                "Failed to retain direct subscription for completed account"
-            );
-        }
-    }
-
-    async fn cleanup_undelegation_tracking_for_completed_account(
-        &self,
-        pubkey: Pubkey,
-    ) {
-        if let Err(err) = self
-            .remote_account_provider
-            .release_subscription_reason_silently_for_delegated_account(
-                &pubkey,
-                SubscriptionReason::UndelegationTracking,
-            )
-            .await
-        {
-            warn!(
-                pubkey = %pubkey,
-                error = %err,
-                "Failed to clean up undelegation tracking for completed account"
-            );
-        }
-    }
-
-    async fn resolve_account_to_clone_from_forwarded_sub_with_unsubscribe(
-        &self,
-        update: ForwardedSubscriptionUpdate,
-        companion_fetch_log_context: &CompanionFetchLogContext,
-    ) -> (
-        Option<AccountSharedData>,
-        Option<DelegationRecord>,
-        DelegationActions,
-    ) {
-        let ForwardedSubscriptionUpdate {
-            pubkey,
-            account,
-            source: _,
-        } = update;
-        let owned_by_delegation_program =
-            account.is_owned_by_delegation_program();
-
-        if let Some(account) = account.fresh_account() {
-            // If the account is owned by the delegation program we need to resolve
-            // its true owner and determine if it is delegated to us
-            if owned_by_delegation_program {
-                let delegation_record_pubkey =
-                    delegation_record_pda_from_delegated_account(&pubkey);
-
-                let acquired_delegation_record_reason = self
-                    .acquire_subscription_reason(
-                        &delegation_record_pubkey,
-                        SubscriptionReason::DelegationRecord,
-                    )
-                    .await
-                    .map(|_| true)
-                    .unwrap_or_else(|err| {
-                        warn!(
-                            pubkey = %delegation_record_pubkey,
-                            error = ?err,
-                            "Failed to acquire delegation record subscription reason"
-                        );
-                        false
-                    });
-
-                match self
-                    .task_to_fetch_with_delegation_record(
-                        pubkey,
-                        ResolvedAccountSharedData::Fresh(account.clone()),
-                        account.remote_slot(),
-                        account.remote_slot(),
-                        AccountFetchContext::subscription_update(
-                            AccountFetchReason::DelegationRecord,
-                        ),
-                    )
-                    .await
-                {
-                    Ok(Ok(AccountWithCompanion {
-                        pubkey,
-                        mut account,
-                        companion_pubkey: delegation_record_pubkey,
-                        companion_account: delegation_record,
-                    })) => {
-                        // We may need to remove temporary subscriptions created
-                        // while resolving this update.
-                        let mut subs_to_remove = Vec::new();
-
-                        subs_to_remove.push(SubscriptionRelease::Pubkey {
-                            pubkey: delegation_record_pubkey,
-                            reason: SubscriptionReason::DirectAccount,
-                        });
-                        if acquired_delegation_record_reason {
-                            subs_to_remove.push(SubscriptionRelease::Pubkey {
-                                pubkey: delegation_record_pubkey,
-                                reason: SubscriptionReason::DelegationRecord,
-                            });
-                        }
-
-                        let account = if let Some(delegation_record) =
-                            delegation_record
-                        {
-                            let delegation_record_with_actions = match self
-                                .parse_delegation_record(
-                                    delegation_record.data(),
-                                    delegation_record_pubkey,
-                                ) {
-                                Ok(x) => Some(x),
-                                Err(err) => {
-                                    error!(
-                                        pubkey = %pubkey,
-                                        error = %err,
-                                        "Failed to parse delegation record"
-                                    );
-                                    None
-                                }
-                            };
-
-                            // If the delegation record is valid we set the owner and delegation
-                            // status on the account
-                            if let Some((
-                                delegation_record,
-                                delegation_actions,
-                            )) = delegation_record_with_actions
-                            {
-                                if tracing::enabled!(tracing::Level::TRACE) {
-                                    let delegation_record_display =
-                                        format!("{:?}", delegation_record);
-                                    trace!(
-                                        pubkey = %pubkey,
-                                        slot = account.remote_slot(),
-                                        owner = %delegation_record.owner,
-                                        deleg_record = %delegation_record_display,
-                                        "Resolving delegated account"
-                                    );
-                                }
-
-                                self.apply_delegation_record_to_account(
-                                    pubkey,
-                                    &mut account,
-                                    &delegation_record,
-                                );
-
-                                // For accounts delegated to us, subscribe to the original owner
-                                // program for undelegation update resilience.
-                                if account.delegated()
-                                    && !self
-                                        .programs_not_to_subscribe
-                                        .contains(&delegation_record.owner)
-                                {
-                                    // Fire-and-forget to avoid blocking subscription updates.
-                                    let provider =
-                                        self.remote_account_provider.clone();
-                                    let owner = delegation_record.owner;
-                                    tokio::spawn(async move {
-                                        if let Err(err) = provider
-                                            .subscribe_program(owner)
-                                            .await
-                                        {
-                                            warn!(
-                                                "Failed to subscribe to owner program {} for account {}: {}",
-                                                owner, pubkey, err
-                                            );
-                                        }
-                                    });
-                                }
-
-                                (
-                                    Some(account.into_account_shared_data()),
-                                    Some(delegation_record),
-                                    delegation_actions.unwrap_or_default(),
-                                )
-                            } else {
-                                // If the delegation record is invalid we cannot clone the account
-                                // since something is corrupt and we wouldn't know what owner to
-                                // use, etc.
-                                (None, None, DelegationActions::default())
-                            }
-                        } else if let Ok(request) =
-                            UndelegationRequest::try_from_bytes_with_discriminator(
-                                account.data(),
-                            )
-                        {
-                            let observed = ObservedUndelegationRequest {
-                                request_pda: pubkey,
-                                delegated_account: request.delegated_account,
-                                expires_at_slot: request.expires_at_slot,
-                                observed_slot: account.remote_slot(),
-                            };
-                            trace!(
-                                request_pda = %observed.request_pda,
-                                delegated_account = %observed.delegated_account,
-                                expires_at_slot = observed.expires_at_slot,
-                                "Observed DLP undelegation request"
-                            );
-                            if let Err(broadcast::error::SendError(observed)) =
-                                self.undelegation_request_sender.send(observed)
-                            {
-                                warn!(
-                                    request_pda = %observed.request_pda,
-                                    delegated_account = %observed.delegated_account,
-                                    observed_slot = observed.observed_slot,
-                                    expires_at_slot = observed.expires_at_slot,
-                                    drop_reason = "no_active_subscribers",
-                                    "Dropped observed DLP undelegation request because no subscribers are active"
-                                );
-                            }
-                            (
-                                Some(account.into_account_shared_data()),
-                                None,
-                                DelegationActions::default(),
-                            )
-                        } else if is_internal_dlp_account_data(account.data()) {
-                            (
-                                Some(account.into_account_shared_data()),
-                                None,
-                                DelegationActions::default(),
-                            )
-                        } else {
-                            trace!(
-                                pubkey = %pubkey,
-                                "Skipping DLP-owned subscription update without delegation record"
-                            );
-                            (None, None, DelegationActions::default())
-                        };
-
-                        if !subs_to_remove.is_empty() {
-                            release_subs(
-                                &self.remote_account_provider,
-                                subs_to_remove,
-                            )
-                            .await;
-                        }
-                        account
-                    }
-                    // In case of errors fetching the delegation record we cannot clone the account
-                    Ok(Err(err)) => {
-                        log_companion_fetch_failure(
-                            companion_fetch_log_context,
-                            delegation_record_pubkey,
-                            ChainlinkCompanionFetchKind::DelegationRecord,
-                            &err,
-                        );
-                        if acquired_delegation_record_reason {
-                            release_subs(
-                                &self.remote_account_provider,
-                                [SubscriptionRelease::Pubkey {
-                                    pubkey: delegation_record_pubkey,
-                                    reason:
-                                        SubscriptionReason::DelegationRecord,
-                                }],
-                            )
-                            .await;
-                        }
-                        (None, None, DelegationActions::default())
-                    }
-                    Err(err) => {
-                        log_companion_fetch_failure(
-                            companion_fetch_log_context,
-                            delegation_record_pubkey,
-                            ChainlinkCompanionFetchKind::DelegationRecord,
-                            &err,
-                        );
-                        if acquired_delegation_record_reason {
-                            release_subs(
-                                &self.remote_account_provider,
-                                [SubscriptionRelease::Pubkey {
-                                    pubkey: delegation_record_pubkey,
-                                    reason:
-                                        SubscriptionReason::DelegationRecord,
-                                }],
-                            )
-                            .await;
-                        }
-                        (None, None, DelegationActions::default())
-                    }
-                }
-            } else {
-                let (account, deleg_record) = self
-                    .maybe_project_ata_from_subscription_update(
-                        pubkey,
-                        account,
-                        companion_fetch_log_context,
-                    )
-                    .await;
-                if let Some((deleg_record, actions)) = deleg_record {
-                    (
-                        Some(account),
-                        Some(deleg_record),
-                        actions.unwrap_or_default(),
-                    )
-                } else {
-                    (Some(account), None, DelegationActions::default())
-                }
-            }
-        } else {
-            // This should not happen since we call this method with sub updates which always hold
-            // a fresh remote account
-            error!(pubkey = %pubkey, account = ?account, "BUG: Received subscription update without fresh account");
-            (None, None, DelegationActions::default())
-        }
-    }
-
-    async fn maybe_build_projected_ata_clone_request_from_subscription_update_with_source(
-        &self,
-        eata_pubkey: Pubkey,
-        eata_account: &AccountSharedData,
-        update_source: SubscriptionSource,
-        deleg_record: Option<&DelegationRecord>,
-        delegation_actions: &DelegationActions,
-        companion_fetch_log_context: &CompanionFetchLogContext,
-    ) -> Option<AccountCloneRequest> {
-        ata_projection::maybe_build_projected_ata_clone_request_from_subscription_update(
-            self,
-            eata_pubkey,
-            eata_account,
-            update_source,
-            deleg_record,
-            delegation_actions,
-            companion_fetch_log_context,
-        )
-        .await
-    }
-
-    #[cfg(test)]
-    fn is_known_empty_eata(&self, eata_pubkey: &Pubkey) -> bool {
-        ata_projection::is_known_empty_eata(self, eata_pubkey)
-    }
-
-    async fn maybe_project_ata_from_subscription_update(
-        &self,
-        ata_pubkey: Pubkey,
-        ata_account: AccountSharedData,
-        companion_fetch_log_context: &CompanionFetchLogContext,
-    ) -> (
-        AccountSharedData,
-        Option<(DelegationRecord, Option<DelegationActions>)>,
-    ) {
-        ata_projection::maybe_project_ata_from_subscription_update(
-            self,
-            ata_pubkey,
-            ata_account,
-            companion_fetch_log_context,
-        )
-        .await
-    }
-
-    /// Parses a delegation record from account data bytes.
-    /// Returns the parsed DelegationRecord, or InvalidDelegationRecord error
-    /// if parsing fails.
-    fn parse_delegation_record(
-        &self,
-        data: &[u8],
-        delegation_record_pubkey: Pubkey,
-    ) -> ChainlinkResult<(DelegationRecord, Option<DelegationActions>)> {
-        delegation::parse_delegation_record(
-            data,
-            delegation_record_pubkey,
-            self.validator_keypair.as_ref(),
-        )
-    }
-
-    /// Applies delegation record settings to an account: sets the owner,
-    /// delegation status, and confined status based on the delegation
-    /// record's authority field.
-    /// Returns commit frequency if account is delegated to us
-    fn apply_delegation_record_to_account(
-        &self,
-        account_pubkey: Pubkey,
-        account: &mut ResolvedAccountSharedData,
-        delegation_record: &DelegationRecord,
-    ) -> Option<u64> {
-        delegation::apply_delegation_record_to_account(
-            self,
-            account_pubkey,
-            account,
-            delegation_record,
-        )
-    }
-
-    /// Returns the pubkey of another validator if account is delegated to them,
-    /// None if delegated to us or delegated to the system program (confined).
-    fn get_delegated_to_other(
-        &self,
-        delegation_record: &DelegationRecord,
-    ) -> Option<Pubkey> {
-        delegation::get_delegated_to_other(self, delegation_record)
-    }
-
-    /// Fetches and parses the delegation record for an account, returning the
-    /// parsed DelegationRecord if found and valid, None otherwise.
-    async fn fetch_and_parse_delegation_record(
-        &self,
-        account_pubkey: Pubkey,
-        min_context_slot: u64,
-        fetch_context: metrics::AccountFetchContext,
-        companion_fetch_log_context: CompanionFetchLogContext,
-    ) -> Option<(DelegationRecord, Option<DelegationActions>)> {
-        delegation::fetch_and_parse_delegation_record(
-            self,
-            account_pubkey,
-            min_context_slot,
-            fetch_context,
-            &companion_fetch_log_context,
-        )
-        .await
-    }
-
-    /// Tries to fetch all accounts in `pubkeys` and clone them into the bank.
-    /// If `mark_empty` is provided, accounts in that list that are
-    /// not found on chain will be added with zero lamports to the bank.
-    ///
-    /// - **pubkeys**: list of accounts to fetch and clone
-    /// - **mark_empty**: optional list of accounts that should be added as empty if not found on
-    ///   chain
-    /// - **slot**: optional slot to use as minimum context slot for the accounts being cloned
-    ///
-    /// NOTE: accounts fetched here have not been found in the bank
-    async fn fetch_and_clone_accounts(
-        &self,
-        pubkeys: &[Pubkey],
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<FetchAndCloneResult> {
-        let accs = match self
-            .fetch_accounts(
-                pubkeys,
-                mark_empty_if_not_found,
-                slot,
-                fetch_context.clone(),
-            )
-            .await
-        {
-            Ok(accs) => accs,
-            Err(err) => {
-                for _ in pubkeys {
-                    metrics::inc_chainlink_clone_accounts_total_with_context(
-                        fetch_context.clone(),
-                        ChainlinkCloneRemoteResult::Failed,
-                        ChainlinkCloneIntent::Unknown,
-                        ChainlinkCloneOutcome::Skipped,
-                    );
-                }
-                return Err(err);
-            }
-        };
-        self.clone_accounts(
-            pubkeys,
-            accs,
-            mark_empty_if_not_found,
-            slot,
-            fetch_context,
-        )
-        .await
-    }
-
-    #[instrument(skip(self, pubkeys, mark_empty_if_not_found), fields(tx_sig = tracing::field::Empty))]
-    async fn fetch_accounts(
-        &self,
-        pubkeys: &[Pubkey],
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<Vec<RemoteAccount>> {
-        if let Some(sig) = fetch_context.signature() {
-            tracing::Span::current().record("tx_sig", sig.to_string());
-        }
-        if tracing::enabled!(tracing::Level::TRACE) {
-            let pubkeys_count = pubkeys.len();
-            trace!(count = pubkeys_count, "Fetching accounts");
-        }
-
-        // Increment fetch counter for testing deduplication (count per account being fetched)
-        self.fetch_count
-            .fetch_add(pubkeys.len() as u64, Ordering::Relaxed);
-
-        // Keep the main account fetch aligned with the freshest observed slot.
-        let min_context_slot = slot.map(|subscription_slot| {
-            subscription_slot.max(self.remote_account_provider.chain_slot())
-        });
-
-        let accs = self
-            .remote_account_provider
-            .try_get_multi(
-                pubkeys,
-                mark_empty_if_not_found,
-                fetch_context,
-                min_context_slot,
-            )
-            .await?;
-
-        if tracing::enabled!(tracing::Level::TRACE) {
-            let accs_count = accs.len();
-            trace!(count = accs_count, "Fetched accounts");
-        }
-        Ok(accs)
-    }
-
-    #[instrument(skip(self, pubkeys, accs, mark_empty_if_not_found), fields(tx_sig = tracing::field::Empty))]
-    async fn clone_accounts(
-        &self,
-        pubkeys: &[Pubkey],
-        accs: Vec<RemoteAccount>,
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<FetchAndCloneResult> {
-        if let Some(sig) = fetch_context.signature() {
-            tracing::Span::current().record("tx_sig", sig.to_string());
-        }
-
-        // Keep resolution fetches aligned with the freshest observed slot.
-        let min_context_slot = slot.map(|subscription_slot| {
-            subscription_slot.max(self.remote_account_provider.chain_slot())
-        });
-
-        let ClassifiedAccounts {
-            not_found,
-            plain,
-            owned_by_deleg,
-            programs,
-            atas,
-        } = pipeline::classify_remote_accounts(accs, pubkeys);
-
-        if tracing::enabled!(tracing::Level::TRACE) {
-            let not_found = not_found
-                .iter()
-                .map(|(pubkey, slot)| (pubkey.to_string(), *slot))
-                .collect::<Vec<_>>();
-            let plain = plain
-                .iter()
-                .map(|p| p.pubkey.to_string())
-                .collect::<Vec<_>>();
-            let owned_by_deleg = owned_by_deleg
-                .iter()
-                .map(|(pubkey, _, slot)| (pubkey.to_string(), *slot))
-                .collect::<Vec<_>>();
-            let programs = programs
-                .iter()
-                .map(|(p, _, _)| p.to_string())
-                .collect::<Vec<_>>();
-            let atas = atas
-                .iter()
-                .map(|(a, _, _, _)| a.to_string())
-                .collect::<Vec<_>>();
-            trace!(
-                "Fetched accounts: \nnot_found:      {not_found:?} \nplain:          {plain:?} \nowned_by_deleg: {owned_by_deleg:?}\nprograms:       {programs:?} \natas:       {atas:?}",
-            );
-        }
-
-        let PartitionedNotFound {
-            clone_as_empty,
-            not_found,
-        } = pipeline::partition_not_found(mark_empty_if_not_found, not_found);
-
-        // For accounts we couldn't find we cannot do anything. We will let code depending
-        // on them to be in the bank fail on its own
-        if !not_found.is_empty() {
-            trace!(
-                "Could not find accounts on chain: {:?}",
-                not_found
-                    .iter()
-                    .map(|(pubkey, slot)| (pubkey.to_string(), *slot))
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        // We mark some accounts as empty if we know that they will never exist on chain
-        if tracing::enabled!(tracing::Level::TRACE)
-            && !clone_as_empty.is_empty()
-        {
-            trace!(
-                "Cloning accounts as empty: {:?}",
-                clone_as_empty
-                    .iter()
-                    .map(|(p, _)| p.to_string())
-                    .collect::<Vec<_>>()
-            );
-        }
-
-        // For potentially delegated accounts we update the owner and delegation state first
-        let ResolvedDelegatedAccounts {
-            mut accounts_to_clone,
-            mut record_subs,
-            missing_delegation_record,
-        } = match pipeline::resolve_delegated_accounts(
-            self,
-            owned_by_deleg,
-            plain,
-            min_context_slot,
-            fetch_context.clone(),
-        )
-        .await
-        {
-            Ok(resolved) => resolved,
-            Err(err) => {
-                release_subs(
-                    &self.remote_account_provider,
-                    pubkeys.iter().copied().map(|pubkey| {
-                        SubscriptionRelease::Pubkey {
-                            pubkey,
-                            reason: SubscriptionReason::DirectAccount,
-                        }
-                    }),
-                )
-                .await;
-                return Err(err);
-            }
-        };
-
-        let ResolvedPrograms {
-            loaded_programs,
-            mut program_data_subs,
-        } = match pipeline::resolve_programs_with_program_data(
-            self,
-            programs,
-            min_context_slot,
-            fetch_context.clone(),
-        )
-        .await
-        {
-            Ok(resolved) => resolved,
-            Err(err) => {
-                let releases = pubkeys
-                    .iter()
-                    .copied()
-                    .map(|pubkey| SubscriptionRelease::Pubkey {
-                        pubkey,
-                        reason: SubscriptionReason::DirectAccount,
-                    })
-                    .chain(record_subs.iter().copied().map(|pubkey| {
-                        SubscriptionRelease::Pubkey {
-                            pubkey,
-                            reason: SubscriptionReason::DirectAccount,
-                        }
-                    }))
-                    .chain(record_subs.iter().copied().map(|pubkey| {
-                        SubscriptionRelease::Pubkey {
-                            pubkey,
-                            reason: SubscriptionReason::DelegationRecord,
-                        }
-                    }))
-                    .collect::<Vec<_>>();
-                release_subs(&self.remote_account_provider, releases).await;
-                return Err(err);
-            }
-        };
-
-        let mut loaded_programs = loaded_programs;
-        let mut all_requested_pubkeys = pubkeys.to_vec();
-        all_requested_pubkeys.extend(record_subs.iter().copied());
-        all_requested_pubkeys.extend(program_data_subs.iter().copied());
-
-        // We will compute subscription cancellations after ATA handling, once accounts_to_clone is finalized
-
-        // Handle ATAs: for each detected ATA, we derive the eATA PDA, subscribe to both,
-        // and, if the ATA is delegated to us and the eATA exists, we clone the eATA data
-        // into the ATA in the bank.
-        // eATA subscriptions are kept implicitly (not tracked for release).
-        let ata_accounts = ata_projection::resolve_ata_with_eata_projection(
-            self,
-            atas,
-            min_context_slot,
-            fetch_context.clone(),
-        )
-        .await;
-        accounts_to_clone.extend(ata_accounts);
-
-        // Ensure all accounts referenced by delegation actions exist and are
-        // cloned before we execute those actions as part of account cloning.
-        let action_dependencies =
-            pipeline::collect_delegation_action_dependencies(
-                &accounts_to_clone,
-            );
-        let action_dependencies_to_fetch = action_dependencies
-            .into_iter()
-            .filter(|dependency| {
-                self.accounts_bank.get_account(dependency).is_none()
-                    && !accounts_to_clone
-                        .iter()
-                        .any(|request| request.pubkey.eq(dependency))
-                    && !loaded_programs
-                        .iter()
-                        .any(|program| program.program_id.eq(dependency))
-            })
-            .collect::<Vec<_>>();
-
-        if !action_dependencies_to_fetch.is_empty() {
-            if tracing::enabled!(tracing::Level::TRACE) {
-                trace!(
-                    dependencies = ?action_dependencies_to_fetch,
-                    "Ensuring delegation action dependencies"
-                );
-            }
-
-            self.fetch_count.fetch_add(
-                action_dependencies_to_fetch.len() as u64,
-                Ordering::Relaxed,
-            );
-            let action_dependency_context = fetch_context
-                .clone()
-                .with_reason(AccountFetchReason::ActionDependencyMissing);
-            let action_dep_accs = self
-                .remote_account_provider
-                .try_get_multi(
-                    &action_dependencies_to_fetch,
-                    None,
-                    action_dependency_context.clone(),
-                    min_context_slot,
-                )
-                .await?;
-            all_requested_pubkeys
-                .extend(action_dependencies_to_fetch.iter().copied());
-
-            let ClassifiedAccounts {
-                not_found,
-                plain,
-                owned_by_deleg,
-                programs,
-                atas,
-            } = pipeline::classify_remote_accounts(
-                action_dep_accs,
-                &action_dependencies_to_fetch,
-            );
-
-            if tracing::enabled!(tracing::Level::TRACE) && !not_found.is_empty()
-            {
-                trace!(
-                    dependencies = ?not_found,
-                    "Delegation action dependencies not found on chain; continuing clone flow"
-                );
-            }
-
-            let ResolvedDelegatedAccounts {
-                accounts_to_clone: action_dep_accounts_to_clone,
-                record_subs: action_dep_record_subs,
-                missing_delegation_record: action_dep_missing_delegation_record,
-            } = match pipeline::resolve_delegated_accounts(
-                self,
-                owned_by_deleg,
-                plain,
-                min_context_slot,
-                action_dependency_context.clone(),
-            )
-            .await
-            {
-                Ok(resolved) => resolved,
-                Err(err) => {
-                    let releases = pipeline::compute_subscription_releases(
-                        &all_requested_pubkeys,
-                        &accounts_to_clone,
-                        &loaded_programs,
-                        record_subs.clone(),
-                        program_data_subs.clone(),
-                    );
-                    release_subs(&self.remote_account_provider, releases).await;
-                    return Err(err);
-                }
-            };
-
-            if !action_dep_missing_delegation_record.is_empty() {
-                let releases = pipeline::compute_subscription_releases(
-                    &all_requested_pubkeys,
-                    &accounts_to_clone,
-                    &loaded_programs,
-                    record_subs
-                        .iter()
-                        .copied()
-                        .chain(action_dep_record_subs.iter().copied())
-                        .collect(),
-                    program_data_subs.clone(),
-                );
-                release_subs(&self.remote_account_provider, releases).await;
-                return Err(ChainlinkError::MissingDelegationActionAccounts(
-                    action_dep_missing_delegation_record
-                        .iter()
-                        .map(|(pubkey, _)| *pubkey)
-                        .collect(),
-                ));
-            }
-
-            all_requested_pubkeys
-                .extend(action_dep_record_subs.iter().copied());
-            record_subs.extend(action_dep_record_subs);
-
-            let ResolvedPrograms {
-                loaded_programs: action_dep_loaded_programs,
-                program_data_subs: action_dep_program_data_subs,
-            } = match pipeline::resolve_programs_with_program_data(
-                self,
-                programs,
-                min_context_slot,
-                action_dependency_context.clone(),
-            )
-            .await
-            {
-                Ok(resolved) => resolved,
-                Err(err) => {
-                    let mut cleanup_accounts_to_clone =
-                        accounts_to_clone.clone();
-                    cleanup_accounts_to_clone
-                        .extend(action_dep_accounts_to_clone.clone());
-                    let releases = pipeline::compute_subscription_releases(
-                        &all_requested_pubkeys,
-                        &cleanup_accounts_to_clone,
-                        &loaded_programs,
-                        record_subs.clone(),
-                        program_data_subs.clone(),
-                    );
-                    release_subs(&self.remote_account_provider, releases).await;
-                    return Err(err);
-                }
-            };
-
-            all_requested_pubkeys
-                .extend(action_dep_program_data_subs.iter().copied());
-            program_data_subs.extend(action_dep_program_data_subs);
-
-            let action_dep_ata_accounts =
-                ata_projection::resolve_ata_with_eata_projection(
-                    self,
-                    atas,
-                    min_context_slot,
-                    action_dependency_context,
-                )
-                .await;
-
-            accounts_to_clone.extend(action_dep_accounts_to_clone);
-            accounts_to_clone.extend(action_dep_ata_accounts);
-            loaded_programs.extend(action_dep_loaded_programs);
-        }
-
-        let releases = pipeline::compute_subscription_releases(
-            &all_requested_pubkeys,
-            &accounts_to_clone,
-            &loaded_programs,
-            record_subs,
-            program_data_subs,
-        );
-
-        pipeline::clone_accounts_and_programs(
-            self,
-            accounts_to_clone,
-            loaded_programs,
-            fetch_context,
-        )
-        .await?;
-
-        release_subs(&self.remote_account_provider, releases).await;
-
-        Ok(FetchAndCloneResult {
-            not_found_on_chain: not_found,
-            missing_delegation_record,
-        })
-    }
-
-    /// Determines if the account finished undelegating on chain.
-    /// If it has finished undelegating, we should refresh it in the bank.
-    /// - **pubkey**: the account pubkey
-    /// - **in_bank**: the account as it exists in the bank
-    ///
-    /// Returns true if the account should be refreshed in the bank
-    async fn should_refresh_undelegating_in_bank_account(
-        &self,
-        pubkey: &Pubkey,
-        in_bank: &AccountSharedData,
-        fetch_context: AccountFetchContext,
-    ) -> RefreshDecision {
-        if in_bank.undelegating() {
-            debug!(
-                pubkey = %pubkey,
-                delegated = in_bank.delegated(),
-                undelegating = in_bank.undelegating(),
-                "Fetching undelegating account"
-            );
-
-            if let Some(eata_pubkey) =
-                ata_projection::derive_eata_pubkey_from_ata_layout(
-                    pubkey, in_bank,
-                )
-            {
-                let undelegating_refresh_context = fetch_context
-                    .clone()
-                    .with_reason(AccountFetchReason::UndelegatingRefresh);
-                let companion_fetch_log_context = CompanionFetchLogContext {
-                    origin: undelegating_refresh_context.clone(),
-                    primary_pubkey: eata_pubkey,
-                    context_slot: self.remote_account_provider.chain_slot(),
-                };
-                let projected_deleg_record = self
-                    .fetch_and_parse_delegation_record(
-                        eata_pubkey,
-                        self.remote_account_provider.chain_slot(),
-                        undelegating_refresh_context,
-                        companion_fetch_log_context,
-                    )
-                    .await;
-                if projected_deleg_record.as_ref().is_some_and(|(record, _)| {
-                    record.owner == EATA_PROGRAM_ID
-                        && record.authority == self.validator_pubkey
-                }) {
-                    debug!(
-                        pubkey = %pubkey,
-                        eata_pubkey = %eata_pubkey,
-                        "Keeping undelegating ATA in bank while companion eATA remains delegated"
-                    );
-                    return RefreshDecision::No;
-                }
-            }
-
-            let undelegating_refresh_context = fetch_context
-                .clone()
-                .with_reason(AccountFetchReason::UndelegatingRefresh);
-            let companion_fetch_log_context = CompanionFetchLogContext {
-                origin: undelegating_refresh_context.clone(),
-                primary_pubkey: *pubkey,
-                context_slot: self.remote_account_provider.chain_slot(),
-            };
-            let deleg_record = self
-                .fetch_and_parse_delegation_record(
-                    *pubkey,
-                    self.remote_account_provider.chain_slot(),
-                    undelegating_refresh_context,
-                    companion_fetch_log_context,
-                )
-                .await;
-
-            if deleg_record.is_none() {
-                // If there is no delegation record then it is possible that the account itself
-                // does not exist either.
-                // In that case we need to refresh it as empty to clear the undelegation state.
-                self.invalidate_mirrored_record(pubkey);
-                return RefreshDecision::YesAndMarkEmptyIfNotFound;
-            }
-
-            let delegated_on_chain =
-                deleg_record.as_ref().is_some_and(|(dr, _)| {
-                    dr.authority.eq(&self.validator_pubkey)
-                        || dr.authority.eq(&Pubkey::default())
-                });
-            let deleg_record = deleg_record.map(|el| el.0);
-            if !account_still_undelegating_on_chain(
-                pubkey,
-                delegated_on_chain,
-                in_bank.remote_slot(),
-                deleg_record,
-                &self.validator_pubkey,
-            ) {
-                debug!(
-                    "Account {pubkey} marked as undelegating will be overridden since undelegation completed"
-                );
-                self.invalidate_mirrored_record(pubkey);
-                return RefreshDecision::Yes;
-            }
-        } else if in_bank.owner().eq(&dlp_api::id()) {
-            debug!(
-                "Account {pubkey} owned by deleg program not marked as undelegating"
-            );
-        }
-        RefreshDecision::No
-    }
-
-    /// Fetch and clone accounts with request deduplication to avoid parallel fetches of the same account.
-    /// This method implements the new logic where:
-    /// 1. Check synchronously if account is in bank, return immediately if found
-    /// 2. If account is pending, add to pending requests and await
-    /// 3. Create pending entries and fetch via RemoteAccountProvider
-    /// 4. Once fetched, clone into bank and respond to all pending requests
-    /// 5. Clear pending requests for that account
-    ///
-    /// Note: since we fetch each account only once in parallel, we also avoid fetching
-    /// the same delegation record in parallel.
-    #[instrument(skip(self, pubkeys, mark_empty_if_not_found))]
-    pub async fn fetch_and_clone_accounts_with_dedup(
-        &self,
-        pubkeys: &[Pubkey],
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-    ) -> ChainlinkResult<FetchAndCloneResult> {
-        self.fetch_and_clone_accounts_with_dedup_forced_refresh(
-            pubkeys,
-            mark_empty_if_not_found,
-            slot,
-            fetch_context,
-            &HashSet::new(),
-            None,
-        )
-        .await
-    }
-
-    async fn fetch_and_clone_accounts_with_dedup_forced_refresh(
-        &self,
-        pubkeys: &[Pubkey],
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        slot: Option<u64>,
-        fetch_context: AccountFetchContext,
-        force_refresh_pubkeys: &HashSet<Pubkey>,
-        confirmed_redelegation: Option<(Pubkey, u64)>,
-    ) -> ChainlinkResult<FetchAndCloneResult> {
-        // We cannot clone blacklisted accounts, thus either they are already
-        // in the bank (e.g. native programs) or they don't exist and the transaction
-        // will fail later
-        let mut pubkeys = pubkeys
-            .iter()
-            .filter(|p| !self.blacklisted_accounts.contains(p))
-            .collect::<Vec<_>>();
-        if tracing::enabled!(tracing::Level::TRACE) {
-            let count = pubkeys.len();
-            trace!(count, "Fetching and cloning accounts with dedup");
-        }
-
-        let mut in_bank = HashSet::new();
-        let mut extra_mark_empty = vec![];
-        let mut bank_hit_no_fetch_non_undelegating_count = 0_u64;
-        let mut bank_hit_no_fetch_undelegating_still_valid_count = 0_u64;
-        let mut bank_hit_no_fetch_undelegating_timeout_count = 0_u64;
-        let mut bank_hit_undelegating_refresh_required_count = 0_u64;
-        let mut bank_miss_remote_required_count = 0_u64;
-        let mut forced_refresh_remote_required_count = 0_u64;
-
-        // Phase 1: Sync bank check — separate undelegating accounts
-        // (which need async RPC) from non-undelegating (handled
-        // synchronously). A forced refresh may replace a locally undelegating
-        // account only when its caller supplies a matching newer delegation
-        // generation. Action dependencies never do; the locked runtime guard
-        // rejects the action if that dependency remains unsafe.
-        let mut undelegating_checks: Vec<(Pubkey, AccountSharedData)> = vec![];
-        for pubkey in pubkeys.iter() {
-            if force_refresh_pubkeys.contains(*pubkey) {
-                if let Some(account_in_bank) =
-                    self.accounts_bank.get_account(pubkey)
-                {
-                    if account_in_bank.undelegating() {
-                        let can_replace_undelegating = confirmed_redelegation
-                            .is_some_and(
-                                |(confirmed_pubkey, delegation_slot)| {
-                                    confirmed_pubkey == **pubkey
-                                        && delegation_slot
-                                            > account_in_bank.remote_slot()
-                                },
-                            );
-                        if !can_replace_undelegating {
-                            bank_hit_no_fetch_undelegating_still_valid_count +=
-                                1;
-                            in_bank.insert(**pubkey);
-                            continue;
-                        }
-                    }
-                }
-                forced_refresh_remote_required_count += 1;
-                continue;
-            }
-            if let Some(account_in_bank) =
-                self.accounts_bank.get_account(pubkey)
-            {
-                if account_in_bank.undelegating() {
-                    undelegating_checks.push((**pubkey, account_in_bank));
-                } else {
-                    if account_in_bank.owner().eq(&dlp_api::id()) {
-                        debug!(
-                            pubkey = %pubkey,
-                            "Account owned by deleg program not marked as undelegating"
-                        );
-                    }
-                    if tracing::enabled!(tracing::Level::TRACE) {
-                        let delegated = account_in_bank.delegated();
-                        let owner = account_in_bank.owner();
-                        trace!(
-                            pubkey = %pubkey,
-                            undelegating = false,
-                            delegated,
-                            owner = %owner,
-                            "Account found in bank in valid state, no fetch needed"
-                        );
-                    }
-                    bank_hit_no_fetch_non_undelegating_count += 1;
-                    in_bank.insert(**pubkey);
-                }
-            } else {
-                bank_miss_remote_required_count += 1;
-            }
-        }
-
-        // Phase 2: Parallel undelegation checks via JoinSet
-        if !undelegating_checks.is_empty() {
-            let mut join_set = JoinSet::new();
-            for (pubkey, account_in_bank) in undelegating_checks {
-                let this = self.clone();
-                let fetch_context = fetch_context.clone();
-                join_set.spawn(async move {
-                    let decision = match tokio::time::timeout(
-                        Duration::from_secs(5),
-                        this.should_refresh_undelegating_in_bank_account(
-                            &pubkey,
-                            &account_in_bank,
-                            fetch_context,
-                        ),
-                    )
-                    .await
-                    {
-                        Ok(decision) => decision,
-                        Err(_timeout) => {
-                            warn!(
-                                pubkey = %pubkey,
-                                "Timeout checking if account is still undelegating after 5 seconds"
-                            );
-                            return (pubkey, None);
-                        }
-                    };
-                    (pubkey, Some(decision))
-                });
-            }
-
-            for (pubkey, decision) in join_set.join_all().await {
-                match decision {
-                    Some(
-                        decision @ (RefreshDecision::Yes
-                        | RefreshDecision::YesAndMarkEmptyIfNotFound),
-                    ) => {
-                        debug!(
-                            pubkey = %pubkey,
-                            "Account completed undelegation which was missed and is fetched again"
-                        );
-                        bank_hit_undelegating_refresh_required_count += 1;
-                        metrics::inc_unstuck_undelegation_count();
-                        if let RefreshDecision::YesAndMarkEmptyIfNotFound =
-                            decision
-                        {
-                            extra_mark_empty.push(pubkey);
-                        }
-                    }
-                    Some(RefreshDecision::No) => {
-                        if tracing::enabled!(tracing::Level::TRACE) {
-                            trace!(
-                                pubkey = %pubkey,
-                                "Undelegating account still valid, no fetch needed"
-                            );
-                        }
-                        bank_hit_no_fetch_undelegating_still_valid_count += 1;
-                        in_bank.insert(pubkey);
-                    }
-                    None => {
-                        bank_hit_no_fetch_undelegating_timeout_count += 1;
-                        in_bank.insert(pubkey);
-                    }
-                }
-            }
-        }
-        metrics::inc_chainlink_bank_precheck_accounts_with_context(
-            fetch_context.clone(),
-            BankPrecheckOutcome::BankHitNoFetch,
-            BankPrecheckReason::NonUndelegatingPresent,
-            bank_hit_no_fetch_non_undelegating_count,
-        );
-        metrics::inc_chainlink_bank_precheck_accounts_with_context(
-            fetch_context.clone(),
-            BankPrecheckOutcome::BankHitNoFetch,
-            BankPrecheckReason::UndelegatingStillValid,
-            bank_hit_no_fetch_undelegating_still_valid_count,
-        );
-        metrics::inc_chainlink_bank_precheck_accounts_with_context(
-            fetch_context.clone(),
-            BankPrecheckOutcome::BankHitNoFetch,
-            BankPrecheckReason::UndelegatingCheckTimeout,
-            bank_hit_no_fetch_undelegating_timeout_count,
-        );
-        metrics::inc_chainlink_bank_precheck_accounts_with_context(
-            fetch_context
-                .clone()
-                .with_reason(AccountFetchReason::UndelegatingRefresh),
-            BankPrecheckOutcome::BankHitUndelegatingRefreshRequired,
-            BankPrecheckReason::UndelegatingRefresh,
-            bank_hit_undelegating_refresh_required_count,
-        );
-        metrics::inc_chainlink_bank_precheck_accounts_with_context(
-            fetch_context.clone(),
-            BankPrecheckOutcome::BankMissRemoteRequired,
-            BankPrecheckReason::Absent,
-            bank_miss_remote_required_count,
-        );
-        metrics::inc_chainlink_bank_precheck_accounts_with_context(
-            fetch_context.clone(),
-            BankPrecheckOutcome::ForcedRefreshRemoteRequired,
-            BankPrecheckReason::ForcedRefresh,
-            forced_refresh_remote_required_count,
-        );
-        pubkeys.retain(|p| !in_bank.contains(p));
-
-        let mut mark_empty_set = mark_empty_if_not_found
-            .unwrap_or(&[])
-            .iter()
-            .copied()
-            .collect::<HashSet<_>>();
-        mark_empty_set.extend(extra_mark_empty);
-
-        let mut waiters: Vec<PendingWaiter> = vec![];
-        let mut claimed_ops: Vec<ClaimedOperation> = vec![];
-        for pubkey in pubkeys {
-            match self
-                .claim_or_join_owned_operation(*pubkey, fetch_context.clone())
-            {
-                PendingClaim::Created(handles) => {
-                    let PendingHandles {
-                        waiter,
-                        deadline,
-                        cancel,
-                        owner,
-                    } = handles;
-                    let waiter_pubkey = waiter.pubkey();
-                    let Some(owner) = owner else {
-                        cancel.notify_waiters();
-                        finish_pending(
-                            &self.pending_requests,
-                            waiter_pubkey,
-                            waiter.generation(),
-                            PendingTerminal::Failed(PendingFailure::Cancelled),
-                        );
-                        return Err(
-                            ChainlinkError::MissingPendingRequestOwner(
-                                waiter_pubkey,
-                            ),
-                        );
-                    };
-                    claimed_ops.push(ClaimedOperation {
-                        pubkey: waiter_pubkey,
-                        generation: waiter.generation(),
-                        deadline,
-                        cancel,
-                        owner,
-                    });
-                    waiters.push(waiter);
-                }
-                PendingClaim::Joined(handles) => waiters.push(handles.waiter),
-            }
-        }
-        if !claimed_ops.is_empty() {
-            self.spawn_batched_owned_operation(
-                claimed_ops,
-                &mark_empty_set,
-                slot,
-                fetch_context.clone(),
-            );
-        }
-
-        let mut final_result = FetchAndCloneResult {
-            not_found_on_chain: vec![],
-            missing_delegation_record: vec![],
-        };
-        for waiter in waiters {
-            let pubkey = waiter.pubkey();
-            match waiter.wait().await? {
-                PendingTerminal::Success(owner_result) => {
-                    for entry in owner_result.not_found_on_chain {
-                        if entry.0 == pubkey {
-                            final_result.not_found_on_chain.push(entry);
-                        }
-                    }
-                    for entry in owner_result.missing_delegation_record {
-                        if entry.0 == pubkey {
-                            final_result.missing_delegation_record.push(entry);
-                        }
-                    }
-                }
-                PendingTerminal::Failed(failure) => {
-                    return Err(failure.into_chainlink_error(pubkey));
-                }
-            }
-        }
-
-        Ok(final_result)
-    }
-
-    /// Resolves an account with its delegation record, pairing the account
-    /// copy the caller already holds (at `account_slot`) with a mirrored
-    /// record when the pair is snapshot-consistent: the record must be
-    /// proven through `slot` AND unchanged since at or before
-    /// `account_slot`, so both sides describe the same chain state. That
-    /// path performs no RPC at all. Every other outcome runs the batched
-    /// two-account fetch, byte-for-byte the prior path.
-    fn task_to_fetch_with_delegation_record(
-        &self,
-        pubkey: Pubkey,
-        account: ResolvedAccountSharedData,
-        account_slot: u64,
-        slot: u64,
-        fetch_context: AccountFetchContext,
-    ) -> task::JoinHandle<ChainlinkResult<AccountWithCompanion>> {
-        let delegation_record_pubkey =
-            delegation_record_pda_from_delegated_account(&pubkey);
-
-        if let Some(MirrorLookup::Hit {
-            data,
-            slot: record_slot,
-        }) = self
-            .record_mirror
-            .as_ref()
-            .map(|mirror| mirror.get(&delegation_record_pubkey, slot))
-        {
-            use metrics::RecordMirrorLookupOutcome as Outcome;
-            if record_slot > account_slot {
-                // The record changed after the held account copy was taken;
-                // only the batched fetch can produce a consistent pair.
-                metrics::inc_record_mirror_lookup(Outcome::Stale);
-            } else if !is_delegation_record_data(&data) {
-                metrics::inc_record_mirror_lookup(Outcome::ParseFallback);
-            } else {
-                metrics::inc_record_mirror_lookup(Outcome::Hit);
-                trace!(
-                    pubkey = %pubkey,
-                    companion = %delegation_record_pubkey,
-                    slot,
-                    "Pairing held account with mirrored delegation record"
-                );
-                let companion_account = ResolvedAccountSharedData::Fresh(
-                    AccountSharedData::from(Account {
-                        lamports: MIRRORED_RECORD_LAMPORTS,
-                        data,
-                        owner: dlp_api::id(),
-                        executable: false,
-                        rent_epoch: 0,
-                    }),
-                );
-                return task::spawn(async move {
-                    Ok(AccountWithCompanion {
-                        pubkey,
-                        account,
-                        companion_pubkey: delegation_record_pubkey,
-                        companion_account: Some(companion_account),
-                    })
-                });
-            }
-        }
-
-        self.task_to_fetch_with_companion(
-            pubkey,
-            delegation_record_pubkey,
-            slot,
-            fetch_context.with_reason(AccountFetchReason::DelegationRecord),
-            ChainlinkCompanionFetchKind::DelegationRecord,
-        )
-    }
-
-    fn task_to_fetch_with_program_data(
-        &self,
-        pubkey: Pubkey,
-        slot: u64,
-        fetch_context: AccountFetchContext,
-    ) -> task::JoinHandle<ChainlinkResult<AccountWithCompanion>> {
-        let program_data_pubkey =
-            get_loaderv3_get_program_data_address(&pubkey);
-        self.task_to_fetch_with_companion(
-            pubkey,
-            program_data_pubkey,
-            slot,
-            fetch_context.with_reason(AccountFetchReason::ProgramData),
-            ChainlinkCompanionFetchKind::ProgramData,
-        )
-    }
-
-    fn task_to_fetch_with_companion(
-        &self,
-        pubkey: Pubkey,
-        companion_pubkey: Pubkey,
-        slot: u64,
-        fetch_context: AccountFetchContext,
-        companion_fetch_kind: ChainlinkCompanionFetchKind,
-    ) -> task::JoinHandle<ChainlinkResult<AccountWithCompanion>> {
-        let provider = self.remote_account_provider.clone();
-        let bank = self.accounts_bank.clone();
-        let fetch_count = self.fetch_count.clone();
-        task::spawn(async move {
-            trace!(
-                pubkey = %pubkey,
-                companion = %companion_pubkey,
-                slot,
-                "Fetching account with companion"
-            );
-
-            // Increment fetch counter for testing deduplication (2 accounts: pubkey + delegation_record_pubkey)
-            fetch_count.fetch_add(2, Ordering::Relaxed);
-
-            provider
-                .try_get_multi_until_slots_match(
-                    &[pubkey, companion_pubkey],
-                    Some(MatchSlotsConfig {
-                        min_context_slot: Some(slot),
-                        ..MatchSlotsConfig::new(companion_fetch_kind)
-                    }),
-                    fetch_context,
-                )
-                .await
-                .map_err(ChainlinkError::from)
-                .and_then(|accs| {
-                    match accs.as_slice() {
-                        [acc_first, acc_last] => {
-                            Ok((acc_first.clone(), acc_last.clone()))
-                        }
-                        _ => Err(ChainlinkError::UnexpectedAccountCount(format!(
-                            "Expected exactly 2 accounts for pubkey {} and companion {}, got {}",
-                            pubkey,
-                            companion_pubkey,
-                            accs.len()
-                        ))),
-                    }
-                })
-                .and_then(|(acc, deleg)| {
-                    Self::resolve_account_with_companion(
-                        &bank,
-                        pubkey,
-                        companion_pubkey,
-                        acc,
-                        deleg,
-                    )
-                })
-        })
-    }
-
-    fn resolve_account_with_companion(
-        bank: &V,
-        pubkey: Pubkey,
-        companion_pubkey: Pubkey,
-        acc: RemoteAccount,
-        companion: RemoteAccount,
-    ) -> ChainlinkResult<AccountWithCompanion> {
-        use RemoteAccount::*;
-        match (acc, companion) {
-            // Account not found even though we found it previously - this is invalid,
-            // either way we cannot use it now
-            (NotFound(_), NotFound(_)) | (NotFound(_), Found(_)) => {
-                Err(ChainlinkError::ResolvedAccountCouldNoLongerBeFound(pubkey))
-            }
-            (Found(acc), NotFound(_)) => {
-                // Only account found without a companion
-                // In case of delegation record fetch the account is either invalid
-                // or a delegation record itself.
-                // Clone it as is (without changing the owner or flagging as delegated)
-                match acc.account.resolved_account_shared_data(bank) {
-                    Some(account) => Ok(AccountWithCompanion {
-                        pubkey,
-                        account,
-                        companion_pubkey,
-                        companion_account: None,
-                    }),
-                    None => Err(
-                        ChainlinkError::ResolvedAccountCouldNoLongerBeFound(
-                            pubkey,
-                        ),
-                    ),
-                }
-            }
-            (Found(acc), Found(comp)) => {
-                // Found the delegation record, we include it so that the caller can
-                // use it to add metadata to the account and use it for decision making
-                let Some(comp_account) =
-                    comp.account.resolved_account_shared_data(bank)
-                else {
-                    return Err(
-                        ChainlinkError::ResolvedCompanionAccountCouldNoLongerBeFound(
-                            companion_pubkey,
-                        ),
-                    );
-                };
-                let Some(account) =
-                    acc.account.resolved_account_shared_data(bank)
-                else {
-                    return Err(
-                        ChainlinkError::ResolvedAccountCouldNoLongerBeFound(
-                            pubkey,
-                        ),
-                    );
-                };
-                Ok(AccountWithCompanion {
-                    pubkey,
-                    account,
-                    companion_pubkey,
-                    companion_account: Some(comp_account),
-                })
-            }
-        }
-    }
-
-    /// Check if an account is currently being watched (subscribed to) by the
-    /// remote account provider
-    pub fn is_watching(&self, pubkey: &Pubkey) -> bool {
-        self.remote_account_provider.is_watching(pubkey)
-    }
-
-    /// Subscribe to updates for a specific account
-    /// This is typically used when an account is about to be undelegated
-    /// and we need to start watching for changes
-    #[instrument(skip(self))]
-    pub(crate) async fn acquire_subscription_reason(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-    ) -> ChainlinkResult<()> {
-        self.remote_account_provider
-            .acquire_subscription(pubkey, reason)
-            .await
-            .map_err(|err| {
-                ChainlinkError::FailedToSubscribeToAccount(*pubkey, err)
-            })
-    }
-
-    pub(crate) async fn ensure_subscription(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-    ) -> ChainlinkResult<()> {
-        self.remote_account_provider
-            .ensure_subscription(pubkey, reason)
-            .await
-            .map_err(|err| {
-                ChainlinkError::FailedToSubscribeToAccount(*pubkey, err)
-            })
-    }
-
-    #[instrument(skip(self))]
-    pub async fn subscribe_to_account_to_track_undelegation(
-        &self,
-        pubkey: &Pubkey,
-    ) -> ChainlinkResult<()> {
-        trace!(
-            pubkey = %pubkey,
-            reason = ?SubscriptionReason::UndelegationTracking,
-            "Subscribing to account"
-        );
-        // Acquire undelegation tracking ownership before/with local
-        // undelegating visibility; any LRU entry created for this reason is
-        // protected from capacity eviction by the provider's bank-state
-        // predicate and ownership filter.
-        self.acquire_subscription_reason(
-            pubkey,
-            SubscriptionReason::UndelegationTracking,
-        )
-        .await
-    }
-
-    pub fn chain_slot(&self) -> u64 {
-        self.remote_account_provider.chain_slot()
-    }
-
-    pub fn received_updates_count(&self) -> u64 {
-        self.remote_account_provider.received_updates_count()
-    }
-
-    pub(crate) fn promote_accounts(&self, pubkeys: &[&Pubkey]) {
-        self.remote_account_provider.promote_accounts(pubkeys);
-    }
-
-    pub fn try_get_removed_account_rx(
-        &self,
-    ) -> ChainlinkResult<mpsc::Receiver<Pubkey>> {
-        Ok(self.remote_account_provider.try_get_removed_account_rx()?)
-    }
-
-    /// Best-effort airdrop helper: if the account doesn't exist in the bank or has 0 lamports,
-    /// create/overwrite it as a plain system account with the provided lamports using the cloner path.
-    #[instrument(skip(self))]
-    pub async fn airdrop_account_if_empty(
-        &self,
-        pubkey: Pubkey,
-        lamports: u64,
-    ) -> ClonerResult<()> {
-        if lamports == 0 {
-            return Ok(());
-        }
-        let remote_slot =
-            if let Some(acc) = self.accounts_bank.get_account(&pubkey) {
-                if acc.lamports() > 0 {
-                    return Ok(());
-                }
-                acc.remote_slot()
-                    .max(self.remote_account_provider.chain_slot())
-            } else {
-                self.remote_account_provider.chain_slot()
-            };
-        // Build a plain system account with the requested balance
-        let mut account =
-            AccountSharedData::new(lamports, 0, &system_program::id());
-        account.set_remote_slot(remote_slot);
-        debug!(
-            pubkey = %pubkey,
-            lamports,
-            remote_slot,
-            "Auto-airdropping account"
-        );
-        let _sig = self
-            .cloner
-            .clone_account(AccountCloneRequest {
-                pubkey,
-                account,
-                commit_frequency_ms: None,
-                delegation_actions: DelegationActions::default(),
-                delegated_to_other: None,
-                needs_undelegation: false,
-            })
-            .await?;
-        Ok(())
     }
 }
